CopulaeClass(fCopulae)	R Documentation

Bivariate Copulae Class

Description

A collection and description of functions to specify the copula class and to investigate bivariate Frechet copulae.

The class representation and methods are:

fCOPULA 	 representation for an S4 object of class "fCOPULA",

show 	 S4 print method.

Frechet Copulae:

pfrechetCopula 	 computes Frechet copula probability.

Usage


show.fCOPULA(object)

pfrechetCopula(u = 0.5, v = u, type = c("m", "pi", "w"), 
    output = c("vector", "list"))
    

Arguments

object 	 [show] -
an S4 object of class "fCOPULA".

output 	 [*frechetCopula] -
output - a character string specifying how the output should be formatted. By default a vector of the same length as u and v. If specified as "list" then u and v are expected to span a two-dimensional grid as outputted by the function grid2d and the function returns a list with elements $x, y, and z which can be directly used for example by 2D plotting functions.

type 	 [*frechetCopula] -
the type of the Frechet copula. A character string selected from: "m", "pi", or "w".

u, v 	 two numeric values or vectors of the same length at which the copula will be computed. If u is a list then the the $x and $y elements will be used as u and v. If u is a two column matrix then the first column will be used as u and the the second as v.

Details

The function pfrechetCopula returns a numeric matrix of probabilities computed at grid positions u|v. The arguments u and v are two single values or two numeric vectors of the same length. If v is not specified then the same values are taken as for u. Alternatively, u may be given as a two column vector or as a list with two entries as vectors. The first column or entry is taken as u and the second as v.

Value

The print method show returns an S4 object of class "fCOPULA". The object contains the following slots:

@call 	 the function call.

@copula 	 the name of the copula.

@param 	 a list whose elements specify the model parameters of the copula.

@title 	 a character string with the name of the copula. This can be overwritten specifying a user defined input argument.

@description 	 a character string with an optional user defined description. By default just the current date will be returned.

The function pfrechetCopula returns a numeric vector of probabilities. An attribute named "control" is added which returns the name of the Frechet copula.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCopulae.061A-CopulaeClass")

## fCOPULA -
   # getClass("fCOPULA")
   
## pcopula -
   # The Frechet Copula - m:
   pfrechetCopula(0.5)
   pfrechetCopula(0.25, 0.75)
   pfrechetCopula(runif(5))
   pfrechetCopula(grid2d())

[Package fCopulae version 221.10065 Index]