ArchimedeanCopulae(fCopulae)	R Documentation

Bivariate Archimedean Copulae

Description

A collection and description of functions to investigate bivariate Archimedean copulae.

Archimedean Copulae Functions:

rarchmCopula 	 Generates Archimedean copula variates,

parchmCopula 	 computes Archimedean copula probability,

parchmCopula 	 computes Archimedean copula density,

rarchmSlider 	 displays interactive plots of variates,

parchmSlider 	 displays interactive plots of probability,

darchmSlider 	 displays interactive plots of density,

archmTau 	 computes Kendall's tau for Archimedean copulae,

archmRho 	 computes Spearman's rho for Archimedean copulae,

archmTailCoeff 	 computes tail dependence for Archimedean copulae,

archmTailPlot 	 plots tail dependence for Archimedean copulae,

Phi 	 Computes generator Phi, inverse and derivatives,

PhiSlider 	 displays interactively generator function,

Kfunc 	 computes copula density and its inverse,

KfuncSlider 	 displays interactively density function,

archmCopulaSim 	 simulates an Archimedean copula,

archmCopulaFit 	 fits the parameters of an Archimedean copula.

Usage


    
rarchmCopula(n, alpha = NULL, type = 1:22)
parchmCopula(u = 0.5, v = u, alpha = NULL, type = 1:22, output = 
    c("vector", "list"), alternative = FALSE )
darchmCopula(u = 0.5, v = u, alpha = NULL, type = 1:22, output = 
    c("vector", "list"), alternative = FALSE )
   
rarchmSlider(B = 10) 
parchmSlider(type = c("persp", "contour"), B = 10)
darchmSlider(type = c("persp", "contour"), B = 10)
    
archmTau(alpha = NULL, type = 1:22, lower = 1.0e-10)
archmRho(alpha = NULL, type = 1:22, method = c("integrate2d", "adapt"), error = 1.0e-5)
    
archmTailCoeff(alpha = NULL, type = 1:22)
archmTailPlot(alpha = NULL, type = 1:22, tail = c("Upper", "Lower"))
   
Phi(x, alpha = NULL, type = 1:22, inv = FALSE, deriv = c(0,1,2))
PhiSlider()

Kfunc(x, alpha = NULL, type = 1:22, inv = FALSE, lower = 1.0e-8)
KfuncSlider()

archmCopulaSim(n, alpha = NULL, type = 1:22)
archmCopulaFit(u, v = NULL, type = 1:22, ...)


Arguments

alpha 	 [Phi*][*archmCopula] -
the parameter of the Archemedean copula. A numerical value.

alternative 	 [*Copula] -
Should the probability be computed alternatively ...

B 	 [*Slider] -
the maximum slider menu value when the boundary value is infinite. By default this is set to 10.

deriv 	 [Phi] -
an integer value. Should the function itself, deriv=0, or the first deriv=1, or second deriv=2 derivative be evaluated?

error 	 [archmRho] -
the error bound to be achieved by the integrate2d integration formula. A numeric value, by default error=1.0e-5.

inv 	 [Phi][Kfunc] -
a logical flag. Should the inverse function be computed?

lower 	 [invK] -
a numeric value setting the lower bound for the internal root finding function uniroot.
[archmTau] -
a numeric value setting the lower bound for the internal integration function integrate.

method 	 [archmRho] -
a character string that determines which integration should be used, either "integrate2d" or "adapt". If the second method is selected the condtributed R package "adapt" is required.

n 	 [rarchmCopula] -
the number of random deviates to be generated, an integer value.

output 	 [*archmCopula] -
output - a character string specifying how the output should be formatted. By default a vector of the same length as u and v. If specified as "list" then u and v are expected to span a two-dimensional grid as outputted by the function grid2d and the function returns a list with elements $x, y, and z which can be directly used for example by 2D plotting functions.

tail 	 [archmTailPlot] -
a character string, either "Upper" or "Lower" denoting which of the two tails should be displayed. By default the upper tail dependence will be considered.

type 	 [*archmCopula][Phi][Kfunc] -
the type of the Archimedean copula. An integer ranging beween 1 and 22. Alternatively type can be specified by a string value of the number. By default the first copula will be chosen.

u, v 	 [*archmCopula] -
two numeric values or vectors of the same length at which the copula will be computed. If u is a list then the the $x and $y elements will be used as u and v. If u is a two column matrix then the first column will be used as u and the the second as v.

x 	 [Kfunc] -
a numeric value or vector ranging between zero and one.
[Phi] -
a numeric value or vector.

... 	 [archmCopulaFit] -
arguments passed to the optimization function nlminb.

Value

Th functions *Spec return an S4 object of class "fCOPULA". The object contains the following slots:

@call 	 the function call.

@copula 	 the name of the copula.

@param 	 a list whose elements specify the model parameters.

@title 	 a character string with the name of the copula. This can be overwritten specifying a user defined input argument.

@description 	 a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

The function pcopula returns a numeric matrix of probabilities computed at grid positions x|y.

The function parchmCopula returns a numeric matrix with values computed for the Archemedean copula.

The function darchmCopula returns a numeric matrix with values computed for thedensity of the Archemedean copula.

The functions Phi* return a numeric vector with the values computed from the Archemedean generator, its derivatives, or its inverse.

The functions cK and {cKInv} return a numeric vector with the values of the density and inverse for Archimedian copulae.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCopulae.063A-ArchimedeanCopulae")
## Not run: 
## fCOPULA -
   # getClass("fCOPULA")
   
## pcopula -
   # The default Normal Copula:
   # contour(pcopula())
## End(Not run)

[Package fCopulae version 221.10065 Index]