ExtremeValueCopulae(fCopulae)	R Documentation

Bivariate Extreme Value Copulae

Description

A collection and description of functions to investigate bivariate extreme value copulae.

Extreme Value Copulae:

pevCopula 	 computes extreme value copula probability,

devCopula 	 computes extreme value copula density,

Afunc 	 computes dependence function,

AFirstDer 	 computes derivative of dependence function,

ASecondDer 	 computes 2nd derivative of dependence function,

ASlider 	 displays interactively dependence function,

pevSlider 	 displays interactive plots of probability,

devSlider 	 displays interactive plots of density.

Empirical Copulae:

pempiricalCopula 	 computes empirical copula probability,

dempiricalCopula 	 computes empirical copula density.

Usage



pevCopula(u = 0.5, v = u, param = NULL, type = c("gumbel", "galambos", 
    "husler.reiss", "tawn", "bb5"), output = c("vector", "list"), 
    alternative = FALSE )
devCopula(u = 0.5, v = u, param = NULL, type = c("gumbel", "galambos", 
    "husler.reiss", "tawn", "bb5"), output = c("vector", "list"), 
    alternative = FALSE )   

pevSlider(type = c("persp", "contour"), B = 10)
devSlider(type = c("persp", "contour"), B = 10)

Afunc(x, param = NULL, type = c("gumbel", "galambos", "husler.reiss", 
    "tawn", "bb5") )
AfuncFirstDer(x, param = NULL, type = c("gumbel", "galambos", "husler.reiss", 
    "tawn", "bb5"), eps = 1.0e-6 )
AfuncSecondDer(x, param = NULL, type = c("gumbel", "galambos", "husler.reiss", 
    "tawn", "bb5") )
AfuncSlider()

pempiricalCopula(u, v, N = 10)
dempiricalCopula(u, v, N = 10)

Arguments

alternative 	 [*evCopula] -
Should the probability be computed alternatively in a direct way from the probability formula or by default via the dependency function?

B 	 [*Slider] -
the maximum slider menu value when the boundary value is infinite. By default this is set to 10.

eps 	 [AFirstDer] -
data below eps and above 1-eps will be set to the border limits.

N 	 [empiricalCopula] -
... .

output 	 [*evCopula] -
output - a character string specifying how the output should be formatted. By default a vector of the same length as u and v. If specified as "list" then u and v are expected to span a two-dimensional grid as outputted by the function grid2d and the function returns a list with elements $x, y, and z which can be directly used for example by 2D plotting functions.

param 	 [*ev*][A*] -
distribution and copulae parameters. A numeric value or vector of named parameters as required by the copula specified by the variable type. If set to NULL, then the default parameters will be taken.

type 	 [*ev*][Afunc] -
the type of the maximum extreme value copula. A character string selected from: "gumbel", "galambos", "husler.reiss", "tawn", or "bb5".
[evSlider] -
a character string specifying the plot type. Either a perspective plot which is the default or a contour plot with an underlying image plot will be created.

u, v 	 [*evCopula][*archmCopula] -
two numeric values or vectors of the same length at which the copula will be computed. If u is a list then the the $x and $y elements will be used as u and v. If u is a two column matrix then the first column will be used as u and the the second as v.

x 	 [A*] -
a numeric value or vector ranging between zero and one.

Value

Th functions *Spec return an S4 object of class "fCOPULA". The object contains the following slots:

@call 	 the function call.

@copula 	 the name of the copula.

@param 	 a list whose elements specify the model parameters.

@title 	 a character string with the name of the copula. This can be overwritten specifying a user defined input argument.

@description 	 a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

The function pcopula returns a numeric matrix of probabilities computed at grid positions x|y.

The function parchmCopula returns a numeric matrix with values computed for the Archemedean copula.

The function darchmCopula returns a numeric matrix with values computed for thedensity of the Archemedean copula.

The functions Phi* return a numeric vector with the values computed from the Archemedean generator, its derivatives, or its inverse.

The functions cK and {cKInv} return a numeric vector with the values of the density and inverse for Archimedian copulae.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCopulae.E4-ExtremeValueCopulae")

## fCOPULA -
   # getClass("fCOPULA")
   
## pcopula -
   # The default Normal Copula:
   # contour(pcopula())

[Package fCopulae version 221.10065 Index]