CopulaeTools(fCopulae)	R Documentation

fCopulae Tools

Description

Popups the example menu and starts or continues the example and demo programs.

Usage


xmpfCopulae()
xmpCopulae(prompt = "")

Arguments

prompt 	 the string printed when prompting the user for input.

Details

The example in the manual pages may be interactive and ask for input from the user. To achieve this you have to type on the command line: xmpCopulae = readline

Value

xmpfCopulae
Popups the example menu.

xmpCopulae
Nothing, the default, or the the prompt if you have set xmpCopulae = readline on the command prompt.

Author(s)

Diethelm Wuertz for this R-Port.

Examples


## SOURCE("fCopulae.Z1-CopulaeTools")

## Not run: 
## xmpfCopulae -
   # Popup the examples menu:
   xmpfCopulae()
## End(Not run)

[Package fCopulae version 221.10065 Index]