RmetricsUtilities(fCalendar)	R Documentation

Masked Objects and Additional Generic Functions

Description

A collection and description of functions which mask function from R's base installation to generate generic functions and some additional generic functions which may be useful.

Usage


log.default(x, base = exp(1))

outlier.default(x, sd = 5, complement = TRUE, ...)

round.default(x, digits, ...)

sample.default(x, size, replace = FALSE, prob = NULL, ...)

sort.default(x, partial = NULL, na.last = NA, decreasing = FALSE, 
    method = c("shell", "quick"), index.return = FALSE, ...) 

var.default(x, y = NULL, na.rm = FALSE, use)
    
rownames(x) <- value
colnames(x) <- value

as.POSIXlt(x, tz = "")

as.matrix.ts(x)
as.matrix.mts(x)

currentYear
myUnits

MSFT

Arguments

base 	 [log] -
additional argument to the log function.

digits 	 [round] -
an integer indicating the precision to be used.

partial, na.last, decreasing, method, index.return 	 [sort] -
additional arguments to the sort function.

sd, complement 	 [outlier] -
sd - a numeric value of standard deviations, e.g. 5 means that values larger or smaller tahn five times the standard deviation of the series will detected. complement - a logical flag, should the outlier series or its complements be returned ?

size, replace, prob 	 [sample] -
size - is a non-negative integer giving the number of items to choose,
replace - a logical flag. Should sampling be with replacement?
prob - a vector of probability weights for obtaining the elements of the vector being sampled.

tz 	 [as.POSIXlt] -
time zone specification.

value 	 [colnames][rownames] -
additional arguments to the colnames and rownames functions.

x 	 [log][sort][var] -
first argument.
[as.matrix.ts][as.matrix.mts] -
an univariate or multivariate time series object of class "ts" or "mts" which will be transformed into an one-column or multi-column rectangular object of class "matrix".
[as.POSIXlt] -
an object to be converted.

y, na.rm, use 	 [var] -
additional arguments to the var function.

... 	 arguments to be passed.

Details

For details we refer to the original help pages.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fBasics.0A-RmetricsUtilities")

[Package fCalendar version 231.10066 Index]