PortfolioData(fPortfolio)	R Documentation

fPortfolio Data Sets

Description

Data sets used in the examples of the 'fPortfolio' package.

The Van Hedge Fund Indices were one of the first performance benchmarks based on a large and representative sample of hedge fund returns. Published from 1995 until now, the Van Indices reflect the average performance of hedge funds back to 1988. The Indices are produced on a monthly basis. They represent the average performance of hedge funds around the world and therefore tracks the performance of the overall hedge fund universe.

All Index returns are based on hedge funds returns that are net of fees. They are simple non-dollar-weighted averages. Indexes for different time periods may be based on different funds, depending on the hedge funds reporting to VAN at the time. The database, used in construction of the Indices, contains information on approximately 5,300 hedge funds.

The Van Global Hedge Fund Index is reported on a percentage-change basis. The Index was set at an initial value of 1,000 as of its January 1, 1988 inception.

Format

The csv data file consists of 16 columns, the first with the date entries formatted as "%m/%d/%Y", followed by 14 columns with the Index, and the last column with the Global Hedge Fund Index.

The Indexes are:

AggressiveGrowth
DistressedSecurities
EmergingMarkets
FundofFunds
Income
Macro
MarketNeutralArbitrage
MarketNeutralHedging
MarketTiming
Opportunistic
SeveralStrategies
ShortSelling
SpecialSituations
Value
VanGlobalHedgeFundIndex

Source

Van Hedge Fund Advisors International, Inc.
http://hedgefund.com/

Examples


## SOURCE("fPortfolio.109A-PortfolioData")

## -

[Package fPortfolio version 221.10065 Index]