fPortfolioTools(fPortfolio)	R Documentation

fPortfolio Tools

Description

Tools used in the 'fPortfolio' package.

Usage


xmpfPortfolio()
xmpPortfolio(prompt = "")

Arguments

prompt 	 the string printed when prompting the user for input.

Value

xmpfPortfolio
Popups the example menu.

xmpPortfolio
Nothing, the default, or the the prompt if you have set xmpPortfolio = readline on the command prompt.

Note

The example in the manual pages may be interactive and ask for input from the user. To achieve this you have to type on the command line: xmpPortfolio = readline

Author(s)

Diethelm Wuertz for the R-Port.

Examples


## SOURCE("fPortfolio.109B-PortfolioTools")

## Not run: 
## xmpfPortfolio -
   # Popup the examples menu:
   xmpfPortfolio()
## End(Not run)


[Package fPortfolio version 221.10065 Index]