WebDataImport(fImport)	R Documentation

Import Market Data from the Internet

Description

A collection and description of functions to import financial and economic market data from the Internet.

The functions are:

fWEBDATA 	 Represents objects of class WEBDATA,

show|print 	 prins objects of class fWEBDATA.

Usage

           
show.fWEBDATA(object)

Arguments

object 	 an S4 object of class "fWEBDATA".

Value

An S4 object of class fWEBDATA has the following slots:

@call 	 the function call.

@data 	 the data as downloaded formatted as a data.frame.

@param 	 a character vector whose elements contain the values of selected parameters of the argument list.

@title 	 a character string with the name of the download. This can be overwritten specifying a user defined input argument.

@description 	 a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

[Package fImport version 251.70 Index]