fBasicsUtilities(fBasics)	R Documentation

General Rmetrics Utilities

Description

A collection and description of functions which are useful utilites and tools used by Rmetrics.

The plot utility functions are:

characterTable 	 Table of Numerical Equivalents to Latin Characters,

symbolTable 	 Table of plot characters, plot symbols,

colorTable 	 Table of Color Codes and Plot Colors itself,

greyPalette 	 Creates a grey palette like rainbow does for colors.

Usage

 
characterTable(font = 1, cex = 0.7)
symbolTable(font = par('font'), cex = 0.7)
colorTable(cex = 0.7)
greyPalette(n = 64, start = 255-n, end = 255)

Arguments

font, cex 	 an integer value, the number of the font, by default font number 1, the standard font for the characterTable or the current plot character font for the plotcharacterTable. The character size is determined by the numeric value cex, the default size is 0.7.

n, start, end 	 [greyPal] - cr n gives the number of greys to be constructed, start and end span the range of the color palette. By default 64 grey tones equidistant chosen from the color range (191, 191, 191) to (255, 255, 255).

Value

characterTable
displays a table with the characters of the requested font. The character on line "xy" and column "z" of the table has code "\xyz", e.g cat("\126") prints: V for font number 1. These codes can be used as any other characters.

symbolTable
displays a table with the plot characters numbered from 0 to 255.

colorTable
displays a table with the plot colors with the associated color number.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fBasics.1A-fBasicsUtilities")

[Package fBasics version 231.10066 Index]