fCalendarUtilities(fCalendar)	R Documentation

Masked Objects and Additional Generic Functions

Description

A collection and description of functions which mask function from R's base installation to generate generic functions and some additional generic functions which may be useful.

Usage


align.default(x, y, xout, method = "linear", n = 50, rule = 1, f = 0, 
    ties = mean, ...)

log.default(x, base = exp(1))

outlier.default(x, sd = 5, complement = TRUE, ...)

round.default(x, digits, ...)

sample.default(x, size, replace = FALSE, prob = NULL, ...)

sort.default(x, partial = NULL, na.last = NA, decreasing = FALSE, 
    method = c("shell", "quick"), index.return = FALSE, ...) 

var.default(x, y = NULL, na.rm = FALSE, use)
    
rownames(x) <- value
colnames(x) <- value

as.POSIXlt(x, tz = "")

as.matrix.ts(x)
as.matrix.mts(x)

currentYear
myUnits

Arguments

base 	 [log] -
additional argument to the log function.

digits 	 [round] -
an integer indicating the precision to be used.

f 	 [align] -
for method="constant" the value of f takes a number between 0 and 1 inclusive, indicating a compromise between left- and right-continuous step functions. If y0 and y1 are the values to the left and right of the point then the value is y0*(1-f)+y1*f so that f=0 is right-continuous and f=1 is left-continuous.

method 	 [align] -
a character string which specifies the alignment method to be used. Choices are "linear" or "constant".

n 	 [align] -
if xout is not specified, alignment takes place at n equally spaced points spanning the interval range(x).

partial, na.last, decreasing, index.return 	 [sort] -
additional arguments to the sort function.

rule 	 [align] -
an integer describing how alignment is to take place outside the interval range(x). If rule=1 then NA's are returned for such points and if is rule=2, the value at the closest data extreme is used.

sd, complement 	 [outlier] -
sd - a numeric value of standard deviations, e.g. 5 means that values larger or smaller tahn five times the standard deviation of the series will detected. complement - a logical flag, should the outlier series or its complements be returned ?

size, replace, prob 	 [sample] -
size - is a non-negative integer giving the number of items to choose,
replace - a logical flag. Should sampling be with replacement?
prob - a vector of probability weights for obtaining the elements of the vector being sampled.

ties 	 [align] -
handles tied x values. Either a function with a single vector argument returning a single number result or the string "ordered".

tz 	 [as.POSIXlt] -
time zone specification.

value 	 [colnames][rownames] -
additional arguments to the colnames and rownames functions.

x 	 [align] -
x-coordinates of the points to be aligned.
[log][sort][var] -
first argument.
[as.matrix.ts][as.matrix.mts] -
an univariate or multivariate time series object of class "ts" or "mts" which will be transformed into an one-column or multi-column rectangular object of class "matrix".
[as.POSIXlt] -
an object to be converted.

xout 	 [align] -
a set of values specifying where interpolation is to take place.

y, na.rm, use 	 [align] -
y-coordinates of the points to be aligned.
[var] -
additional arguments to the var function.

... 	 arguments to be passed.

Details

For details we refer to the original help pages.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCalendar.1A-fCalendarUtilities")

[Package fCalendar version 231.10066 Index]