CodeTables(fUtilities)	R Documentation

Charaqcter Symbol and Color Tables

Description

A collection and description of functions which are useful utilities and tools in selectings characters, symbols and colors for plotting.

The plot utility functions are:

characterTable 	 Table of Numerical Equivalents to Latin Characters,

symbolTable 	 Table of plot characters, plot symbols,

colorTable 	 Table of Color Codes and Plot Colors itself.

Usage

 
characterTable(font = 1, cex = 0.7)
symbolTable(font = par('font'), cex = 0.7)
colorTable(cex = 0.7)

Arguments

font, cex 	 [*Table] -
an integer value, the number of the font, by default font number 1, the standard font for the characterTable or the current plot character font for the plotcharacterTable. The character size is determined by the numeric value cex, the default size is 0.7.

Value

characterTable
displays a table with the characters of the requested font. The character on line "xy" and column "z" of the table has code "\xyz", e.g cat("\126") prints: V for font number 1. These codes can be used as any other characters.

symbolTable
displays a table with the plot characters numbered from 0 to 255.

colorTable
displays a table with the plot colors with the associated color number.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## Character Table for Font 2:
   characterTable(font = 1)
   
## Default Symbol Table:
   symbolTable()
    
## Color Table:
   colorTable()

[Package fUtilities version 251.70 Index]