DaylightSavingTime(fCalendar)	R Documentation

Daylight Saving Time Rules

Description

A collection and description of functions for 92 cities and regions, which return time zone offsets and daylight saving time rules.

The functions are:

Adelaide Algiers Amsterdam Anchorage Andorra Athens Auckland Bahrain Bangkok Beirut Belfast Belgrade Berlin Bogota Bratislava Brisbane Brussels Bucharest Budapest BuenosAires Cairo Calcutta Caracas Casablanca Cayman Chicago Copenhagen Darwin Denver Detroit Dubai Dublin Eastern Edmonton Frankfurt Helsinki HongKong Honolulu Indianapolis Istanbul Jakarta Jerusalem Johannesburg Kiev KualaLumpur Kuwait Lagos Lisbon Ljubljana London LosAngeles Luxembourg Madrid Manila Melbourne MexicoCity Monaco Montreal Moscow Nairobi Nassau NewYork Nicosia Oslo Pacific Paris Perth Prague Riga Riyadh Rome Seoul Shanghai Singapore Sofia Stockholm Sydney Taipei Tallinn Tehran Tokyo Tunis Vaduz Vancouver Vienna Vilnius Warsaw Winnipeg Zagreb Zurich.

Note

There are currently two synonymes available "Pacific" for Los Angeles and "Eastern" for New York.

Note that we leave the space in all double named cities like New York or Hong Kong and don't use an underscore for it.

All the entries are retrieved from the Ical library which is available under GNU's GPL licence.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCalendar.2B-DaylightSavingTime")

## ...
   # DST Rules for Zurich:
   head(Zurich())
   tail(Zurich())

[Package fCalendar version 231.10066 Index]