PosixStandard(fCalendar)	R Documentation

Summary of Posix Standard

Description

A summary and description how Rmetrics uses the POSIX standard.

The functions from R's base Package are:

strptime 	 Turns charater representations into POSIXlt,

as.POSIXlt 	 Converts into POSIXlt,

ISOdatetime 	 Returns an object of class POSIXct,

ISOdate 	 Returns an object of class POSIXct,

seq 	 Creates sequence of class POSIXct,

as.POSIXct 	 Converts into POSIXct,

Sys.timeDate 	 Returns system time as an object of class 'timeDate',

format 	 Turns POSIXt into character representations,

strftime 	 Represents an alias for format.POSIXlt,

as.character 	 Calls format function,

Sys.time 	 Returns system time,

Sys.timezone 	 Returns system timezone,

print 	 Prints POSIXt objects, generic function,

"-" 	 Subtracts,

time1 lop time2 	 Logical Operation into PSOIXlt,

time1 - time2 	 Returns an obejct of class difftime,

difftime 	 Takes a difference,

as.difftieme 	 Converts into difftime,

round 	 Rounds difftime, generic function,

as integer 	 Converts POSIXt into integer,

as.numeric 	 Converts POSIXt into integer,

as.character 	 Converts POSIXt into character,

as.data.frame 	 Converts POSIXt into data frame,

"[" 	 Subsets a POSIXt object,

c 	 Concatenates POSIXt objects,

min, max, range 	 Minimum, Maximum, Range of POSIXt objects,

sort, rev, uniq 	 Sorts, reverts, makes unique POSIXt objects,

Sys.Date 	 Returns current Date as Date object,

format 	 Turns Date into character representations,

print 	 Prints POSIXt objects, generic function,

seq 	 Creates sequence of class Date.

The function addons are:

isPOSIX 	 Checks for an object of class POSIX.

Usage


isPOSIX(x)

Arguments

x 	 an argument to be checked.

Author(s)

Diethelm Wuertz for this summary.

Examples


## SOURCE("fCalendar.2A-PosixStandard")

## isPOSIX -
   isPOSIX(Sys.time())
   isPOSIX(Sys.timeDate())

[Package fCalendar version 231.10066 Index]