ColorPalettes(fUtilities)	R Documentation

Color and Grey Palettes

Description

A collection and description of functions for using and managing color and grey palettes.

Color and Grey Palettes:

greyPalette 	 Creates a grey palette like rainbow does for colors.

Usage

 
greyPalette(n = 64, start = 255-n, end = 255)

Arguments

n, start, end 	 [greyPal] - cr n gives the number of greys to be constructed, start and end span the range of the color palette. By default 64 grey tones equidistant chosen from the color range (191, 191, 191) to (255, 255, 255).

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## GreyPalette:
   greyPalette()

[Package fUtilities version 251.70 Index]