PlotUtilities(fUtilities)	R Documentation

Explorative Data Analysis

Description

A collection and description of functions for plotting with emphasis on interactive plots.

The plot utility functions are:

interactivePlot 	 a framework for interactive plot displays,

gridVector 	 creates from two vectors x and y all grid points.

Usage


interactivePlot(x, choices = paste("Plot", 1:9), 
    plotFUN = paste("plot.", 1:9, sep = ""), which = "all", ...)
gridVector(x, y)

Arguments

choices 	 [interactivePlot] -
a vector of character strings for the choice menu. By Default "Plot 1" ... "Plot 9" allowing for 9 plots at maximum.

plotFUN 	 [interactivePlot] -
a vector of character strings naming the plot functions. By Default "plot.1" ... "plot.9" allowing for 9 plots at maximum.

which 	 plot selection, which graph should be displayed? If "which" is a character string named "ask" the user is interactively asked which to plot, if a logical vector of length N, those plots which are set TRUE are displayed, if a character string named "all" all plots are displayed.

x, y 	 [gridVector] -
two numeric vector which span the two dimensional grid.

... 	 additional arguments passed to the FUN or plot function.

Author(s)

Diethelm Wuertz for this R port.

Examples


## Test Plot Function:
   testPlot = function(x, which = "all", ...) {   
     # Plot Function and Addons:
     plot.1 <<- function(x, ...) plot(x, ...)      
     plot.2 <<- function(x, ...) acf(x, ...)
     plot.3 <<- function(x, ...) hist(x, ...)      
     plot.4 <<- function(x, ...) qqnorm(x, ...)
     # Plot:
     interactivePlot(x,
       choices = c("Series Plot", "ACF", "Histogram", "QQ Plot"),
       plotFUN = c("plot.1", "plot.2", "plot.3", "plot.4"),
       which = which, ...)       
     # Return Value:
     invisible()
   }  
   # Plot:
   par(mfrow = c(2, 2), cex = 0.7)
   testPlot(rnorm(500))
      
   # Try:
   # par(mfrow = c(1,1))
   # testPlot(rnorm(500), which = "ask")
        
## Grid Vector - 
   gridVector(1:5, 1:5)

[Package fUtilities version 251.70 Index]