fBasicsData(fBasics)	R Documentation

fBasics Data Sets

Description

A collection and description of data sets used in the examples. Included are data files for high frequency FX data, time and sales data for financial futures, and market returns for selected stock and market indexes. Included are also tables for the finite sample Jarque-Bera test.

The data sets are:

audusd.csv 	 Reuters Tick-by-Tick AUDUSD rates 1997-10,

usdthb.csv 	 Reuters Tick-by-Tick USDTHB rates 1997,

usddem30u.csv 	 Reuter 30 min USDDEM rate in upsilon time,

usdchf.csv 	 Reuters 30 min USDCHF Rates 199604-200103,

fdax9710.csv 	 Minute-by-Minute DAX Futures Prices for 1997-10*,

fdax97m.csv 	 Minutely Time and Sales DAX Futures for 1997,

bmwres.csv 	 Daily log Returns of German BMW Stock Proces,

nyse.csv 	 Daily Values of the NYSE Composite Index,

nyseres.csv 	 Daily log Returns of the NYSE Composite Index,

jbLM 	 Table for the Jarque Bera Lagrange Multiplier test,

jbALM 	 Table for the JB Augmented LM finite sample test,

PhiStable 	 Table of Contours for Stable Parameter Estimation.

The CIA Factbook:

ciaFactbook.R 	 Macroeconomic data from the CIA Factbook.

The Tables from World Federarion of Exchanges , WFE:

wfe1.csv 	 Table 1, market capitalization of domestic companies,

wfe2.csv 	 Table 2, total number of companies with shares listed,

wfe3.csv 	 Table 3, total value of share trading,

wfe4.csv 	 Table 4, market value of bonds listed,

wfe5.csv 	 Table 5, total value of bond trading,

wfe6.csv 	 Table 6, price earning ratio an gross dividend yield.

*The file fdax97m.csv is too large and therefore not part of the fBasics distribution. Please contact inf@rmetrics.org.

Details

High Frequency Data for the AUDUSD and USDTHB:

audusd and usdthb archive high frequency exchange rates for the Australian / US Dollar exchange rate in October 1997 and exchange rates for the US Dollar / Thailand Bhat exchange rate in June 1997: A comma delimited CSV file with 6 columns. The first column, named XDATE, contains date/time entries in ISO-8601 format as [CCYYMMDDhhmm], the second column, named DELAY, gives the delay in minutes between the time stamp of Reuter's data record and arrival time at the local database, the third column named CONTRIBUTOR is Reuter's identification, a 4 character code, the fourth and fifth column, named BID and ASK are the bid and ask price quotations, and finally the sixth column, named FLAG, is not used and has zeros as entries.

30 Minutes Data for the USDDEM Rate in Upsilon time:

usddem30u archives 30 min USDDEM bid and ask for the US Dollar German Mark exchange rate ranging from 1992-10-01 00:09 until 1997-05-30 21:49. A comma delimited CSV file with 3 columns. The first column, named "%Y%m%d%H%M", contains date/time entries in ISO-8601 format as [YYYYMMDDhhmm], the second column, named BID, gives the bid prices, and the third column named ASK, gives the ask prices retrieved from the Reuter's data records.

30 Minutes Data for the USDCHF Rate:

usdchf archives 30 min USDCHF midprices for the US Dollar Swiss Franc exchange rate ranging from 1996-04-01 until 2001-03-30. A comma delimited CSV file with 2 columns. The first column, named "%Y%m%d%H%M", contains date/time entries in ISO-8601 format as [YYYYMMDDhhmm], the second column, named USDCHF, gives the prices retrieved from the Reuter's data records.

DAX Futures Data:

fdax9710 archives returns of minute-by-minute prices for Dax Futures in October 1997: A comma delimited CSV file with 2 columns. The first column, named XDATE, contains date/time entries in ISO-8601 format as [CCYYMMDDhhmm], the second column, named FDAX, gives an averaged price of the Dax Futures, i.e. the mean of all volume weighted time and sales within the same minute.
fdax97m archives returns for minute-by-minute prices for Dax Futures in 1997: A comma delimited CSV file with 2 columns. The first column, named XDATE, contains date/time entries in ISO-8601 format as [CCYYMMDDhhmm], the second column, named FDAX, gives a minutely averaged price during opening hours of the exchange, i.e. the mean of all volume weighted time and sales within the same minute.

Log Returns for BMW Shares and NYSE Composite Index:

bmwres and nyseres archive log returns of the German BMW stock listed in the German DAX30 and log returns of the NYSE Composite Index, both on a daily trading day time scale just numbering the log returns: A one column CSV file with column names BMW or NYSERES, respectively. The entries are the differences of the logarithmic prices on two succeeding trading days. nyse contains two column data recors, the date and the NYSE Composite Index.

Jarque Bera Normality Test:

jbLM and jbALM are finite sample tables for the Jarque Bera Lagrange multiplier and augmented Lagrange multiplier normality test. The columns denote the sample sizes and the row the probabilities.

Stable Parameter Estimation:

.PhiStable is a list object containing two tables for the estimation of the parameters of a stable distribution using McCullochs approach.

Source

audusd
usdthb
usdthb
The data were collected by D. Wuertz and R. Schnidrig from the Reuter's data feed.

fdax9710
fdax97m
The data were extracted from time and sales data records from the Frankfurt Futures Exchange.

bmwres
The data were published in the EVIS software package.

nyse
nyseres
The data were downloaded from the web site of the New York Stock Exchange and the residuals were calculated as logarithmic price differences.
http://www.nyse.com.

jbLM
jbALM
Monte Carlo Simulations by D. Wuertz and H.G. Katzgraber.

Examples


## SOURCE("fBasics.2C-fBasicsData")

##  plot -
    data(nyseres)
    x = as.ts(nyseres)
    par(mfrow = c(2, 1), cex = 0.7)
    plot(100*x, type = "l", col = "steelblue4",
      main = "NYSE Composite Index")
    grid()
    plot(cumsum(x), type = "l", col = "steelblue4",
      main = "Cumulated NYSE Index")
    grid()

[Package fBasics version 221.10066 Index]