ZivotWangData(fBasics)	R Documentation

fBasics Data Sets from MoFiTS

Description

A collection and description of public data sets used in the examples of the book 'Modeling Financial Time Series with S-Plus' written by E. Zivot and J. Wang.

The data sets are:

CPI.dat 	 Seasonally adjusted US Consumer Price Index,

DowJones30.csv 	 Dow Jones Industrial Average Index,

ford.s.csv 	 Daily returns of the Ford stock,

highFreq3M.df 	 3M High Frequency Stock Market Data,

hp.s.csv 	 Daily returns of the HP stock,

IP.dat 	 Seasonally Adjusted US Industrial Production Index,

lexrates.dat 	 Spot and Forward Exchange Rate Data,

msft.dat 	 Open, High, Low, Close of Microsoft Stocks,

shiller.dat 	 Robert Shiller's Monthly Economic Data Set,

shiller.annual 	 Robert Shiller's Monthly Economic Data Set,

singleIndex.dat.csv 	 Microsoft Stocks and SP500 Index Data,

varex.ts.csv 	 Real Stock Returns and Output Growth Data,

yhoo.df 	 Yahoo Stock Information.

Details

Dow Jones Industrial Average:

The file DowJones30.csv contains closing prices for the 30 stocks represented in the Dow Jones Industrial Average Index.
Data are downloadable and can be updated from Yahoo's web site.

Microsoft Stocks and SP500 Index Data:

The file singleIndex.dat.csv contains the monthly closing prices for Microsoft Corporation and the SP 500 index.

Open, High, Low, Close of Microsoft Stocks:

The file msft.dat.csv contains data representing the open, high, low, close and volume information for Microsoft stocks.
Data are downloadable and can be updated from Yahoo's web site.

IP and CPI Index Data:

The file IP.dat.csv contains data representing seasonally adjusted US Industrial Production Index and the file CPI.dat.csv contains data representing seasonally adjusted US Consumer Price Index.
Data are downloadable and can be updated from Economagics's web site.

Ford and HP Stock Returns:

The files ford.s.csv and hp.s.csv contain data representing 2000 daily stock returns for the Ford and HP shares traded at NYSE. The time series span the period from Feburary 2, 1984, to December 31, 1991.
Data are downloadable and can be updated from Yahoo's web site.

3M high Frequency Data:

The file highFreq3M.df.csv holds Tsay's data for 3M. Date information is expressed as the day of the month and the number of seconds from midnight Data is for December 1999. Columns are: day - integer representing the trading day of the month, sec - trade.time integer representing the trading time recorded as the number of seconds from midnight, price transaction price in dollars. Downloadable from:
http://www.gsb.uchicago.edu/fac/ruey.tsay/teaching/fts

Spot and Forward Exchange Rate Data:

The file lexrates.dat.csv holds log spot exchange and forward exchange rates between USCNS, USCNF - USD and CAD, Canadian Dollar, USDMS, USDMF - USD and DEM, Deutsche Mark, USFRS, USFRF - USD and FFR, French Franc, USILS, USJYF - USD and JPY, Italian Lira, and USUKS, USUKF - USD and GBP, Japanese Yen.
Source: Thompson Financial, formerly Datastream, see also: Zivot, E. (2000) Cointegration and forward and spot exchange rate regressions, Journal of International Money and Finance, 19, 387–401 and 785–812.

Robert Shillers Monthlly and Annual Economic Data:

The files shiller.dat.csv and shiller.annual.csv hold data used in the book "Irrational Exuberance" by Robert Shiller. The data are price - monthly nominal US SP stock market prices, dividend - nominal SP Composite Index dividends, earnings - nominal SP Composite Index earnings, cpi - US Consumer Price Indexes, real.price - real US stock market prices, real.dividend - real SP Composite Index dividends, real.earnings - real SP Composite Index earnings, pe.10 - price-earnings ratios, dp.ratio - dividend-price ratios, dp.yield - dividend-price yield. The last two are only listed in shiller.annual. The series start January 1871 and end on March 2001. Data are available from:
Shiller, R.J. (1989) Market Volatility, MIT Press.
Shiller, R.J. (2001) Irrational Exuberance, Broadway Books.

Yahoo Stock Information:

The file yhoo.df.csv contains data representing daily transaction information of Yahoo stock, with the following six columns: Date, Open, High, Low, Close, Volume.
Data are downloadable and can be updated from Yahoo's web site.

Real Stock Returns and Output Growth Data:

The file varex.ts.csv contains real stock returns and output growth data. The column MARKET.REAL lists continuously compounded real returns on the SP500 index, the column RF.REAL lists real interest rates of 30-day US Treasury Bills, the column INF lists continuously compounded growth rate of US CPI, and the column IPG lists continuously compounded growth rate of US industrial production. The data are monthly starting December 1959 and ending December 2000.
Data are downloadable and can be updated from Economagic's web site.

Examples


## SOURCE("fBasics.2D-ZivotWangData")

## Not run: 
## DowJones30 - 
   require(fCalendar)
   data(DowJones30)
   class(DowJones30)
   DowJones30.ts = as.timeSeries(DowJones30)
   class(DowJones30.ts)
   head(DowJones30.ts)
## End(Not run)

[Package fBasics version 221.10066 Index]