TimeDateFinCenter(fCalendar)	R Documentation

timeDate Financial Centers

Description

A collection and description of functions to use financial centers in contex with timeDate objects.

The functions for Financial Centers are:

rulesFinCenter 	 Returns DST rules for a financial center,

listFinCenter 	 Lists all supported financial centers.

Usage


myFinCenter

rulesFinCenter(FinCenter = myFinCenter)
listFinCenter(pattern = ".*")

Arguments

FinCenter 	 a character with the the location of the financial center named as "continent/city".

pattern 	 a pattern character string as required by the grep function.

Details

Financial Centers

There are two functions concerned with the financial centers. The first, rulesFinCenter, lists the daylight saving rules for a selected financial center, and the second, listFinCenter, lists all centers available in the database. There is no dependency on the POSIX implementation of your operating system because all time zone and day light saving time information is stored locally in ASCII files. It is important to say, that the TZ environment variable must set to "GMT" in your system environment that there are no conflicts with the POSIX time zone management.

Through the definition of financial centers it becomes possible to introduce in the future a specification structure for financial centers, which includes further information for a center like holiday calendars, currency and interest rate conventions or many others.

Value

rulesFinCenter
returns a printed list of DST rules.

listFinCenter
returns a list of time zones available in the database.

Author(s)

Diethelm Wuertz for the Rmetrics R-port; some improvements: Martin Maechler.

See Also

timeDateCoercion, timeDateSpecDates

We also recommend to inspect the help pages for the POSIX time and date class, Dates, and the help pages from the contributed R packages chron and date.

Examples


## myFinCenter - the global setting currently used:
   myFinCenter

## Other Financial Centers:
   listFinCenter("Asia/")
   listFinCenter("^A") # all beginning with "A"
   listFinCenter("^[^A]") # all *not* beginning with "A"
   listFinCenter(".*/L")  # cities with L*

   stopifnot(identical(listFinCenter(), ## 'A' and 'not A' == everything:
        union(listFinCenter("^A"),
            listFinCenter("^[^A]"))))

[Package fCalendar version 251.70 Index]