SkewnessKurtosis(fUtilities)	R Documentation

Skdewness and Kurtosis

Description

A collection and description of functions to compute skewness and kurtosis.

The functions are:

skewness 	 returns value of skewness,

kurtosis 	 returns value of kurtosis.

Usage


skewness(x, ...)
## Default S3 method:
skewness(x, na.rm = FALSE, method = c("moment", "fisher"), ...)
## S3 method for class 'data.frame':
skewness(x, ...)
## S3 method for class 'POSIXct':
skewness(x, ...)
## S3 method for class 'POSIXlt':
skewness(x, ...)

kurtosis(x, ...)
## Default S3 method:
kurtosis(x, na.rm = FALSE, method = c("excess", "moment", "fisher"), ...)
## S3 method for class 'data.frame':
kurtosis(x, ...)
## S3 method for class 'POSIXct':
kurtosis(x, ...)
## S3 method for class 'POSIXlt':
kurtosis(x, ...)

Arguments

na.rm 	 a logical. Should missing values be removed?

method 	 [skewness][kurtosis] -
a character string which specifies the method of computation. These are either "moment" or "fisher", kurtosis allows in addition for "excess". If "excess" is selected, then the value of the kurtosis is computed by the "moment" method and a value of 3 will be subtracted. The "moment" method is based on the definitions of skewness and kurtosis for distributions; these forms should be used when resampling (bootstrap or jackknife). The "fisher" method correspond to the usual "unbiased" definition of sample variance, although in the case of skewness and kurtosis exact unbiasedness is not possible.

x 	 a numeric vector.
[basicStats] -
allows also a matrix, data.frame or timeSeries object as input. In this case only the first column of data will be considered and a a warning will be printed.

... 	 arguments to be passed.

Value

skewness
kurtosis
return the value of the statistics, a numeric value. An attribute which reports the used method is added.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples

   
## mean -
## var -
   # Mean, Variance:
   r = rnorm(100)
   mean(r)
   var(r)
   
## skewness -
## kurtosis -
   # Skewness, Kurtosis:
   skewness(r)
   kurtosis(r)   

[Package fUtilities version 251.70 Index]