ColumnRowStatistics(fUtilities)	R Documentation

Column and Row Statistics

Description

A collection and description of functions to compute column and row statistical properties of financial and economic time series data.

The functions are:

rowStats 	 calculates row statistics,

colStats 	 calculates column statistics,

rowAvgs|rowSums 	 calculates row sums,

colAvgs|colSums 	 calculates column sums,

rowAvgs|rowMeans 	 calculates row means,

colAvgs|colMeans 	 calculates column means,

rowVars 	 calculates row variances,

colVars 	 calculates column variances,

rowStdevs 	 calculates row standard deviations,

colStdevs 	 calculates column standard deviations,

rowSkewness 	 calculates row skewness,

colSkewness 	 calculates column skewness,

rowKurtosis 	 calculates row kurtosis,

colKurtosis 	 calculates column kurtosis,

rowCumsums 	 calculates row cumulated Sums,

colCumsums 	 calculates column cumulated Sums.

Usage


rowStats(x, FUN, na.rm = FALSE, ...) 
## S3 method for class 'timeSeries':
rowSums(x, na.rm = FALSE, ...)
## S3 method for class 'timeSeries':
rowMeans(x, na.rm = FALSE, ...)
rowAvgs(x, na.rm = FALSE, ...)
rowVars(x, na.rm = FALSE, ...)
rowStdevs(x, na.rm = FALSE, ...)
rowSkewness(x, na.rm = FALSE, ...)
rowKurtosis(x, na.rm = FALSE, ...)
rowCumsums(x, na.rm = FALSE, ...)

colStats(x, FUN, na.rm = FALSE, ...) 
## S3 method for class 'timeSeries':
colSums(x, na.rm = FALSE, ...)
## S3 method for class 'timeSeries':
colMeans(x, na.rm = FALSE, ...)
colAvgs(x, na.rm = FALSE, ...)
colVars(x, na.rm = FALSE, ...)
colStdevs(x, na.rm = FALSE, ...)
colSkewness(x, na.rm = FALSE, ...)
colKurtosis(x, na.rm = FALSE, ...)
colCumsums(x, na.rm = FALSE, ...)

Arguments

FUN 	 [colStats][rowStats -
the statistical function to be applied.

na.rm 	 a logical. Should missing values be removed?

x 	 a rectangular object which can be transformed into a matrix by the function as.matrix. Note, this allows also a data.frame or timeSeries object as input.

... 	 arguments to be passed.

Value

rowStats
rowAvgs
rowVars
rowStdevs
rowSkewness
rowKurtosis
rowCumsum
compute sample statistics by column. Missing values can be handled.

colStats
colAvgs
colVars
colStdevs,
colSkewness
colKurtosis
colCumsum
compute sample statistics by column. Missing values can be handled.

Note

R's-base package contains a function colMeans with an additional argument dim=1. Therefore, the function used here to compute column means (averages) is named colAvgs.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## Simulated Monthly Return Data:
   x = matrix(rnorm(24), ncol = 2)
     
## colStats -
   colStats(x, FUN = mean)  

[Package fUtilities version 251.70 Index]