HeavisideFunction(fUtilities)	R Documentation

Haviside and Related Functions

Description

A collection and description of functions which compute the Heaviside and related functions. These include the sign function, the delta function, the boxcar function, and the ramp function.

The functions are:

Heaviside 	 Computes Heaviside unit step function,

Sign 	 Just another signum function,

Delta 	 Computes delta function,

Boxcar 	 Computes boxcar function,

Ramp 	 Computes ramp function.

Usage


Heaviside(x, a = 0)
Sign(x, a = 0)
Delta(x, a = 0)
Boxcar(x, a = 0.5)
Ramp(x, a = 0)

Arguments

a 	 a numeric value, the location of the break.

x 	 a numeric vector.

Details

The Heaviside step function Heaviside is 1 for x>a, 1/2 for x=a, and 0 for x<a.

The Sign function Sign is 1 for x>a, 0 for x=a, and -1 for x<a.

The delta function Delta is defined as: Delta(x) = d/dx H(x-a).

The boxcar function Boxcar is defined as: Boxcar(x) = H(x+a) - H(x-a).

The ramp function is defined as: Ramp(x)= (x-a)*H(x-a).

Value

returns the function values of the selected function.

Note

The Heaviside function is used in the implementation of the skew Normal, Student-t, and Generalized Error distributions, distributions functions which play an important role in modelling GARCH processes.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

References

Weisstein W. (2004); http://mathworld.wolfram.com/HeavisideStepFunction.html, Mathworld.

See Also

GarchDistribution, GarchDistributionFits.

Examples


## Heaviside -
   x = sort(round(c(-1, -0.5, 0, 0.5, 1, 5*rnorm(5)), 2))
   h = Heaviside(x)
    
## Sign -
   s = Sign(x)
   
## Delta -
   d = Delta(x)
   
## Boxcar -
   Pi = Boxcar(x)

## Ramp - 
   r = Ramp(x)
   cbind(x = x, Step = h, Signum = s, Delta = d, Pi = Pi, R = r)        

[Package fUtilities version 251.70 Index]