TimeSeriesClass(fCalendar)	R Documentation

timeSeries Class and Methods

Description

A collection and description of functions and methods dealing with regular and irregular 'timeSeries' objects. Dates and times are implemented as 'timeDate' objects. Included are functions and methods for the generation and representation of 'timeSeries' objects, and for mathematical operations.

Functions to generate and modify 'timeSeries' objects:

timeSeries 	 Creates a 'timeSeries' object from scratch,

readSeries 	 Reads a 'timeSeries' from a spreadsheet file,

returnSeries 	 Computes returns from a 'timeSeries' object,

applySeries 	 Applies a function to margins of a 'timeSeries',

orderStatistics 	 Computes order statistic of a 'timeSeries'.

Data Slot and classification of 'timeSeries' objects:

seriesData 	 Extracts data slot from a 'timeSeries',

isUnivariate 	 Tests if a 'timeSeries' object is univariate,

isMultivariate 	 Tests if a 'timeSeries' object is multivariate.

Functions to print and plot 'timeSeries' objects:

print.timeSeries 	 S3: Print method for a 'timeSeries' object,

plot.timeSeries 	 S3: Plot method for a 'timeSeries' object,

lines.timeSeries 	 S3: Lines method for a 'timeSeries' object,

points.timeSeries 	 S3: Points method for a 'timeSeries' object.

Special Functions for daily 'timeSeries' objects:

dummyDailySeries 	 Creates a dummy daily 'timeSeries' object,

alignDailySeries 	 Aligns a daily 'timeSeries' to new positions,

ohlcDailyPlot 	 Plots open high low close bar chart.

Usage


timeSeries(data, charvec, units = NULL, format = NULL, zone = myFinCenter, 
    FinCenter = myFinCenter, recordIDs = data.frame(), title = NULL, 
    documentation = NULL, ...)   
readSeries(file, header = TRUE, sep = ";", zone = myFinCenter, 
    FinCenter = myFinCenter, title = NULL, documentation = NULL, ...)
returnSeries(x, type = c("continuous", "discrete"), percentage = FALSE, 
    trim = TRUE, digits = 4, units = NULL)    
applySeries(x, from = NULL, to = NULL, by = c("monthly", "quarterly"), 
    FUN = colAvgs, units = NULL, ...)
    
orderStatistics(x)

seriesData(object)
isUnivariate(x)
isMultivariate(x)

## S3 method for class 'timeSeries':
print(x, ...)
## S3 method for class 'timeSeries':
plot(x, reference.grid = TRUE, lty = 1, ...) 
## S3 method for class 'timeSeries':
lines(x, ...)
## S3 method for class 'timeSeries':
points(x, ...)

dummyDailySeries(x)
alignDailySeries(x, method = c("before", "after", "interp", "fillNA"), 
    include.weekends = FALSE, units = NULL)
ohlcDailyPlot(x, volume = TRUE, colOrder = c(1:5), units = 1e6, 
    xlab = c("Date", "Date"), ylab = c("Price", "Volume"), 
    main = c("O-H-L-C", "Volume"), grid.nx = 7, grid.lty = "solid", ...) 

Arguments

by 	 [applySeries] -
a character either "monthly" or "quarterly". The default value is "monthly". Only operative when both arguments from and to have ther default values NULL. In this case the function FUN will be applied to monthly or quarterly periods.

charvec 	 a character vector of dates and times.

colOrder 	 [ohlcDailyPlot] -
an integer vector which gives the order of the prices and the volume in the input object. By default the following order of columns from 1 to 5 is assumed: Open, high, low, close, and volume.

data 	 a data.frame or a matrix object of numeric data.

digits 	 [returnSeries] -
an integer value. The number of digits to be printed in the output.

documentation 	 optional documentation string, or a vector of character strings.

file 	 the filename of a spreadsheet data set from which to import the data records.

FinCenter 	 a character with the the location of the financial center named as "continent/city".

header 	 a logical value indicating whether the file contains the names of the variables as its first line. If missing, the value is determined from the file format: 'header' is set to 'TRUE' if and only if the first row contains one fewer field than the number of columns.

format 	 the format specification of the input character vector,
[as.timeSeries] -
a character string with the format in POSIX notation to be passed to the time series object.

from, to 	 starting date and end date, to must be after from.

FUN 	 [applySeries] -
a function to use for aggregation, by default colAvgs.

grid.lty, grid.nx 	 [ohlcDailyPlot] -
The type of grid line and the number of grid lines used in the plot.

include.weekends 	 [alignDailySeries] -
a logical value. Should weekend dates be included or removed from the series.

lty 	 [plot] -
an integer value selecting the type of the line plot.

main 	 [ohlcDailyPlot] -
a character string to title the price and volume plot.

method 	 [alignDailySeries] -
the method to be used for the alignment. A character string, one of "before", use the data from the row whose position is just before the unmatched position, or "after", use the data from the row whose position is just after the unmatched position, or "linear", interpolate linearly between "before" and "after".

object 	 [is][seriesData][seriesPositions][summary] - an object of class timeSeries.

percentage 	 [returnSeries] -
a logical value. By default FALSE, if TRUE the series will be expressed in percentage changes.

recordIDs 	 a data frame which can be used for record identification.

reference.grid 	 [plot] -
a logical value. Should a reference grid be added to the plot?.

sep 	 [readSeries] -
the field seperator used in the spreadsheet file to separate columns.

title 	 an optional title string, if not specified the inputs data name is deparsed.

trim 	 [diffSeries][returnSeries] -
a logical value. By default TRUE, the first missing observation in the return series will be removed.

type 	 [returnSeries] -
a character string specifying how to compute the returns. Valid choices are: continuous and discrete. For the default type="continuous", the returns are calculated as the logarithmic differences, otherwise if type="discrete", the returns are calculated as percentage changes.

units 	 [applySeries][lag][allignDailySeries][returnSeries][mergeSeries] -
an optional character string, which allows to overwrite the current column names of a timeSeries object. By default NULL which means that the column names are selected automatically.
[ohlcDailyPlot] -
a numeric value, specifying in which multiples the volume should be referenced on the plot labels. By default 1e6, i.e. in units of 1 Million.

volume 	 [ohlcDailyPlot] -
a logigical value. Should a volume plot added to the OHLC Plot. By default TRUE.

x 	 [as] -
a matrix type object to be converted.
[as.vector][as.matrix][as.data.frame] -
[applySeries] -
[cut][end][mergeSeries][plot][print][rev][start] -
an object of class timeSeries.

xlab, ylab 	 [ohlcDailyPlot] -
two string vectors to name the x and y axis of the price and volume plot.

zone 	 the time zone or financial center where the data were recorded.

... 	 arguments passed to other methods.

Details

Generation of Time Series Objects:

We have defined a timeSeries class which is in many aspects similar to the S-Plus class with the same name, but has also some important differences. The class has seven Slots, the 'Data' slot which holds the time series data in matrix form, the 'position' slot which holds the time/date as a character vector, the 'format' and 'FinCenter' slots which are the same as for the 'timeDate' object, the 'units' slot which holds the column names of the data matrix, and a 'title' and a 'documentation' slot which hold descriptive character strings. Date and time is managed in the same way as for timeDate objects.

Value

timeSeries
readSeries
returnSeries
applySeries
return a S4 object of class timeSeries.

orderStatistics
returns ...

seriesData

extracts the @Data slot from a timeSeries object. Thus, seriesData returns an object of class matrix.

isUnivariate
isMultivariate

returns ...

plot
lines
points
print
plot and print methods for an object of class timeSeries.

Note

These functions were written for Rmetrics users using R and Rmetrics under Microsoft's Windows operating system where timze zones, daylight saving times and holiday calendars are insuffeciently supported.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCalendar.4A-TimeSeriesClass")

## data - Data Frame:
   # Load Microsoft Data:
   myFinCenter = "GMT"
   dfMSFT = as.timeSeries(MSFT)
   head(dfMSFT)

## timeSeries - Conversion:
   # Create a timeSeries Objec - The Direct Way ...
   Close = dfMSFT[, 5]
   head(Close)
   # From Scratch ...
   data = as.matrix(dfMSFT[, 4])
   charvec = rownames(dfMSFT)
   Close = timeSeries(data, charvec, units = "Close")
   head(Close)
   c(start(Close), end(Close))
    
## cutSeries -    
   # Cut out April Data from 2001:
   tsApril01 = cut(Close, "2001-04-01", "2001-04-30") 
   tsApril01

## returnSeries -  
   # Compute Returns:
   args(returnSeries)
   # Continuous Returns:
   returnSeries(tsApril01)
   # Discrete Returns:
   returnSeries(tsApril01, type = "discrete")
   # Don't trim:
   returnSeries(tsApril01, trim = FALSE)
   # Use Percentage Values:
   returnSeries(tsApril01, percentage = TRUE, trim = FALSE)
    
## alignDailySeries -     
   # Align with NA:
   args(alignDailySeries)
   tsRet = returnSeries(tsApril01, trim = TRUE)
   GoodFriday(2001)   
   EasterMonday(2001) 
   alignDailySeries(tsRet, method = "fillNA", include.weekends = FALSE)
   alignDailySeries(tsRet, method = "fillNA", include.weekends = TRUE)
    
## alignDailySeries - 
   # Interpolate:
   tsRet
   alignDailySeries(tsRet, method = "interp", include.weekend = FALSE)
   alignDailySeries(tsRet, method = "interp", include.weekend = TRUE)

## applySeries -   
   # Aggregate weekly:
   GoodFriday(2001)
   to = timeSequence(from = "2001-04-11", length.out = 3, by = "week") 
   from = to - 6*24*3600
   from
   to
   applySeries(tsRet, from, to, FUN = sum)

[Package fCalendar version 231.10066 Index]