TimeSeriesPositions(fCalendar)	R Documentation

timeSeries Class and Methods

Description

A collection and description of functions and methods extracting and modifying positions on 'timeSeries' objects.

The functions and methods for the Generation of 'timeSeries' Objects are:

seriesPositions 	 Extracts positions slot from a 'timeSeries',

newPositions<- 	 Modifies positions of a 'timeSeries' object,

sample.timeSeries 	 Resamples a 'timeSeries' object in time,

sort.timeSeries 	 Sorts reverts a 'timeSeries' object in time,

rev.timeSeries 	 Reverts a 'timeSeries' object in time,

start.timeSeries 	 Extracts start date of a 'timeSeries' object,

end.timeSeries 	 Extracts end date of a 'timeSeries' object.

Usage


seriesPositions(object)
newPositions(object) <- value

## S3 method for class 'timeSeries':
start(x, ...)
## S3 method for class 'timeSeries':
end(x, ...)

## S3 method for class 'timeSeries':
sample(x, ...)
## S3 method for class 'timeSeries':
sort(x, ...)
## S3 method for class 'timeSeries':
rev(x)

Arguments

method 	 [alignDailySeries] -
the method to be used for the alignment. A character string, one of "before", use the data from the row whose position is just before the unmatched position, or "after", use the data from the row whose position is just after the unmatched position, or "linear", interpolate linearly between "before" and "after".

object 	 [is][seriesData][seriesPositions][summary] - an object of class timeSeries.

value 	 a valid value for that component of newPositions(x), i.e. an object of class "timeDate" with appropriate length.

x 	 [as] -
a matrix type object to be converted.
[as.vector][as.matrix][as.data.frame] -
[applySeries] -
[cut][end][mergeSeries][plot][print][rev][start] -
an object of class timeSeries.

... 	 arguments passed to other methods.

Value

timeSeries
read.timeSeries
as.timeSeries
return a S4 object of class timeSeries.

seriesData
seriesPositions
extract the @Data and @position slots from a timeSeries object. Thus, seriesData returns an object of class matrix, and seriesPositions returns an object of class timeDate.

is.timeSeries
returns TRUE or FALSE depending on whether its argument is of timeSeries type or not.

aggregateSeries
applySeries
cutSeries
mergeSeries
returnSeries
revSeries
return a S4 object of class timeSeries.

end, start
return a S4 object of class timedate. These are the start and end dates of a timeSeries object.

as.vector
as.matrix
as.data.frame
these are methods which convert a S4 object of class timeSeries either to a vector, a matrix or to a data frame.

plot
lines
points
print
plot and print methods for an object of class timeSeries. Note that the plot function requires the packages its and Hmisc.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCalendar.4C-TimeSeriesPositions")
    
## Create dummy timeSeries:     
   X = timeSeries(matrix(rnorm(24), 12), timeCalendar())
   
## seriesPositions -
   seriesPositions(X)

[Package fCalendar version 231.10066 Index]