GarchOxInterface(fSeries)	R Documentation

R Interface for Garch Ox

Description

A collection and description of functions to fit the parameters of an univariate time series to GARCH models interfacing the G@RCH Ox Package.

The family of GARCH time series models includes the following processes:

1 	 garch 	 generalized AR conditional heteroskedastic models,

2 	 egarch 	 exponential GARCH models,

3 	 aparch 	 asymmetretic power ARCH models.

Usage


garchOxFit(formula, data, cond.dist = c("gaussian", "t", "ged", "skewed-t"), 
    include.mean = TRUE, trace = TRUE, control = list(), title = NULL,
    description = NULL)
    
## S3 method for class 'garchOx':
print(x, digits, ...)
## S3 method for class 'garchOx':
summary(object, ...)
## S3 method for class 'garchOx':
plot(x, ...)

Arguments

cond.dist 	 a character string describing the distribution of innovations. By default the optimization is based on gaussian log likelihood parameter optimization denoted by "gaussian". Alternatively, a Student-t "t", a generalized error "sged", or a skewed Student-t "skewed-t" can be chosen.

control 	 a list of additional control parameters:
truncation - the number of truncation points,by default 100,
xscale - should the time series be scaled by the standard deviation ?

data 	 an optional timeSeries or data frame object containing the variables in the model. If not found in data, the variables are taken from environment(formula), typically the environment from which armaFit is called. If data is an univariate series, then the series is converted into a numeric vector and the name of the response in the formula will be neglected.

description 	 a character string which allows for a brief description.

digits 	 the number of digits to be printed.

formula 	 [garchFit] -
formula object describing the mean and variance equation of the ARMA-GARCH/APARCH model. A pure GARCH(1,1) model is selected when e.g. formula=~garch(1,1). To specify for example an ARMA(2,1)-APARCH(1,1) use formula = ~arma(2,1)+apaarch(1,1).

include.mean 	 should the mean be included? By default TRUE.

object 	 an object of class garchOx as returned from the function code{garchOxFit}.

title 	 a character string which allows for a project title.

trace 	 a logical flag. Should the estimation process be ttraced? By default TRUE.

x 	 an object of class garchOx as returned from the function garchOxFit.

... 	 additional arguments to be passed to the print, summary, and plot methods.

Details

Ox Interface:
The function garchOxFit interfaces a subset of the functionality of the G@ARCH 4.0 Package written in Ox. G@RCH 4.0 is one of the most sophisticated packages for modelling univariate GARCH processes including GARCH, EGARCH, GJR, APARCH, IGARCH, FIGARCH, FIEGARCH, FIAPARCH and HYGARCH models. Parameters can be estimated by approximate (Quasi-) maximum likelihood methods under four assumptions: normal, Student-t, GED or skewed Student-t errors.

About Ox:
Ox (tm) is an object-oriented matrix language with a comprehensive mathematical and statistical function library. Many packages were written for Ox including software mainly for econometric modelling. The Ox packages for time series analysis and forecasting, Arfima, Doornik and Ooms [2003], Garch, Laurent and Peters [2005], and State Space Modelling, Koopman, Shepard and Doornik [1998], are especially worth to note. Since most of the R-users wan't to change to another Statistical Computing environment, we made selected parts of the G@RCH Ox software available for them through an R-Interface. What you have to do, is to read carefully the "Ox citation and copyright" rules and if you agree and fullfill the conditions, then download the OxConsole Software together with the "OxGarch" Package, currently G@RCH 4.0. If you are not qualified for a free license, order your copy from Timberlake Consultants. We recommend to install the "Setup.exe" under the path "C:\Ox\Ox3" and to unzip the OxGarch Package in the directory "C:\Ox\Ox3\Packages". An Update to Ox4 has not yet be done.

Distribution:
Ox and G@RCH are distributed by Timberlake Consultants Ltd. Timberlake Consultants can be contacted through the following web site: www.timberlake.co.uk.

Installation of the Interface:
In addition you have to copy the file "GarchOxModelling.ox" (which is the interface written especially for Rmetrics) from the "fSeries/ox/" directory to the Ox library directory "C:\Ox\lib".

Ox Citation and Copyright Rules:
Ox and all its components are copyright of Jurgen A. Doornik. The Console (command line) versions may be used freely for academic research and teaching purposes only. Commercial users and others who do not qualify for the free version must purchase the Windows version of Ox and GiveWin with documentation, regardless of which version they use (so even when only using Ox on Linux or Unix). Ox should be cited whenever it is used. Refer to the two references given below. Note, failure to cite the use of Ox in published work may result in loss of the right to use the free version, and an invoice at the full commercial price. Ox is available from Timberlake Consultants. The Ox syntax is public, and you may do with your own Ox code whatever you wish, including the file "GarchOxModelling.ox".

Work to do:
Note, only a small part of the functionalities are interfaced until now to R. But, principally it would be possible to interface also other functionalities offered by the Ox Garch Package. This work is left to the Ox/Rmetrics user.

Author(s)

Jurgen A. Doormik for the Ox Environment, www.doornik.com,
Sebastian Laurent for the Ox Garch package, www.garch.org,
Diethelm Wuertz for R's Ox Garch interface.

References

Doornik J.A. (2002), Object-Oriented Matrix Programming Using Ox, London, 3rd ed.: Timberlake Consultants Press and Oxford: www.doornik.com.

Doornik J.A., Ooms M. (2003), Computational Aspects of Maximum Likelihood Estimation of Autoregressive Fractionally Integrated Moving Average Models, Computational Statistics and Data Analysis 42, 333–348.

Koopman J.S., Shepard N., Doornik J.A. (1999), Statistical Algorithms for Models in State Space using SsfPack 2.2, Econometrics Journal 2, 113–166.

Laurent S., Peters J.P. (2002); G@RCH 2.2: An Ox Package for Estimating and Forecasting Various ARCH Models, Journal of Economic Surveys 16, 447–485.

Laurent S., Peters J.P., [2005], G@RCH 4.0, Estimating and Forecasting ARCH Models, Timberlake Consultants, www.timberlake.co.uk

Examples


## SOURCE("fSeries1.4D-GarchOxModelling")

## Not run: 
## garchOxFit -
   
   # Load Benchmark Data Set:
   data(dem2gbp)
   x = dem2gbp[, 1]
   
   # Fit GARCH(1,1):
   garchOxFit(formula.mean = ~arma(0,0), formula.var = ~garch(1,1))
## End(Not run)

[Package fSeries version 231.10066 Index]