HolidayDates(fCalendar)	R Documentation

Public and Ecclesiastical Holidays

Description

A collection and description of functions and methods dealing with holiday dates in the G7 countries and Switzerland.

Usage


Septuagesima(year)
Quinquagesima(year)
AshWednesday(year)
PalmSunday(year)
GoodFriday(year)
Easter(year)
EasterSunday(year)
EasterMonday(year)
RogationSunday(year)
Ascension(year)
Pentecost(year)
PentecostMonday(year)
TrinitySunday(year)
CorpusChristi(year)
ChristTheKing(year)
Advent1st(year)
Advent2nd(year)
Advent3rd(year)
Advent4th(year)
ChristmasEve(year)
ChristmasDay(year)
BoxingDay(year)
NewYearsDay(year)
SolemnityOfMary(year)
Epiphany(year)
PresentationOfLord(year)
Annunciation(year)
TransfigurationOfLord(year)
AssumptionOfMary(year)
BirthOfVirginMary(year)
CelebrationOfHolyCross(year)
MassOfArchangels(year)
AllSaints(year)
AllSouls(year)
LaborDay(year)
CHBerchtoldsDay(year)
CHSechselaeuten(year)
CHAscension(year)
CHConfederationDay(year)
CHKnabenschiessen(year)
GBMayDay(year)
GBBankHoliday(year)
GBSummerBankHoliday(year)
GBMilleniumDay(year)
DEAscension(year)
DECorpusChristi(year)
DEGermanUnity(year)
DEChristmasEve(year)
DENewYearsEve(year)
FRFetDeLaVictoire1945(year)
FRAscension(year)
FRBastilleDay(year)
FRAssumptionVirginMary(year)
FRAllSaints(year)
FRArmisticeDay(year)
ITEpiphany(year)
ITLiberationDay(year)
ITAssumptionOfVirginMary(year)
ITAllSaints(year)
ITStAmrose(year)
ITImmaculateConception(year)
USDecorationMemorialDay(year)
USPresidentsDay(year)
USNewYearsDay(year)
USInaugurationDay(year)
USMLKingsBirthday(year)
USLincolnsBirthday(year)
USWashingtonsBirthday(year)
USMemorialDay(year)
USIndependenceDay(year)
USLaborDay(year)
USColumbusDay(year)
USElectionDay(year)
USVeteransDay(year)
USThanksgivingDay(year)
USChristmasDay(year)
USCPulaskisBirthday(year)
USGoodFriday(year)
CAVictoriaDay(year)
CACanadaDay(year)
CACivicProvincialHoliday(year)
CALabourDay(year)
CAThanksgivingDay(year)
CaRemembranceDay(year)
JPNewYearsDay(year)
JPGantan(year)
JPBankHolidayJan2(year)
JPBankHolidayJan3(year)
JPComingOfAgeDay(year)
JPSeijinNoHi(year)
JPNatFoundationDay(year)
JPKenkokuKinenNoHi(year)
JPGreeneryDay(year)
JPMidoriNoHi(year)
JPConstitutionDay(year)
JPKenpouKinenBi(year)
JPNationHoliday(year)
JPKokuminNoKyujitu(year)
JPChildrensDay(year)
JPKodomoNoHi(year)
JPMarineDay(year)
JPUmiNoHi(year)
JPRespectForTheAgedDay(year)
JPKeirouNOhi(year)
JPAutumnalEquinox(year)
JPShuubunNoHi(year)
JPHealthandSportsDay(year)
JPTaiikuNoHi(year)
JPNationalCultureDay(year)
JPBunkaNoHi(year)
JPThanksgivingDay(year)
JPKinrouKanshaNoHi(year)
JPEmperorsBirthday(year)
JPTennouTanjyouBi(year)
JPBankHolidayDec31(year)

Arguments

year 	 an integer value or vector of year numbers including the century. These are integers of the form CCYY, e.g. 2000.

Value

Returns an ISO-8601 formatted calendar date CCYYMMDD, an object of class "sdate" which represents an integer value or vector.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## CHSechselaeuten -
   # Sechselaeuten a half Day Bank Holiday in Switzerland
   CHSechselaeuten(2000:2010)
   CHSechselaeuten(currentYear)
   
## German Unification Day:
   DEGermanUnity(currentYear)

[Package fCalendar version 251.70 Index]