CalendarData(fCalendar)	R Documentation

fCalendar Data Sets

Description

A collection and description of data sets used in the the 'fCalendar' package. This data collection holds the timezone data file with all available 'Ical' time zone names.

The timezone datafile is:

timezones.db 	 file with all available Ical time zone names.

Details

Calendar Data:

holidays.db.R is the database of world wide holidays with informations collected from several sources on the internet.
timezones.csv is a list of timezones information extracted from Olsen's database.

Source

The database was obtained from Downloadable from
ftp://elsie.nci.nih.gov/pub/

Examples


## SOURCE("fCalendar.5A-CalendarData")


[Package fCalendar version 221.10066 Index]