BivariateInterpolation(fUtilities)	R Documentation

Bivariate Interpolation

Description

A collection and description of functions which allow to interpolate and smooth bivariate irregular data sets including linear interpolation, Akima spline interpolation, and kriging.

Gridded Interpolation Functions:

linearInterp 	 performs linear spline interpolation,

akimaInterp 	 performs Akima spline interpolation,

krigeInterp 	 performs krige interpolation.

Pointwise Interpolation Functions:

linearInterpp 	 performs linear spline interpolation,

akimaInterpp 	 performs Akima spline interpolation

Usage

 
linearInterp(x, y = NULL, z = NULL, gridPoints = 21,
    xo = seq(min(x), max(x), length = gridPoints),
    yo = seq(min(y), max(y), length = gridPoints))
linearInterpp(x, y = NULL, z = NULL, xo, yo)
     
akimaInterp(x, y = NULL, z = NULL, gridPoints = 21,
    xo = seq(min(x), max(x), length = gridPoints),
    yo = seq(min(y), max(y), length = gridPoints), extrap = FALSE)
akimaInterpp(x, y = NULL, z = NULL, xo, yo, extrap = FALSE)

krigeInterp(x, y = NULL, z = NULL, gridPoints = 21,
    xo = seq(min(x), max(x), length = gridPoints),
    yo = seq(min(y), max(y), length = gridPoints), 
    extrap = FALSE, polDegree = 6)

Arguments

x, y, z 	 [griddata] -
x and y are two numeric vectors of grid pounts and z is a numeric matrix or any other rectangular object which can be transformed by the function as.matrix into a matrix object.
[akimaInterp][krigeInterp] -
either three numeric vectors of equal length or if y and z are NULL, a list with entries x, y, z, or named data frame with x in the first, y in the second, and z in the third column.
[persp][contour] -
x is an pbject of class gridData.

gridPoints 	 an integer value specifying the number of grid points in x and y direction.

xo, yo 	 [*Interp] -
two numeric vectors of data points spanning the grid.
[*Interpp] -
two numeric vectors of data points building pairs for pointwise interpolation.

extrap 	 a logical, if TRUE then the data points are extrapolated.

polDegree 	 [krige] -
the polynomial krige degree, an integer ranging between 1 and 6.

Value

linearInterp -
akimaInterp -
krigeInterp -
A list with at least three entries, x, y and z. Note, that the returned values, can be directly used by the persp.gridData() and contour.gridData 3D plotting methods.
linearInterpp -
akimaInterpp -
A data.frame with columns "x", "y", and "z".

Note

The function krigeInterp requires loading of the R package spatial.

Author(s)

Diethelm Wuertz for the Rmetrics R-port,
H. Akima for the Fortran Code of the Akima spline interpolation routine.

Examples

 
## akimaInterp -
   # Akima Interpolation:    
   set.seed(1953)
   x = runif(999) - 0.5
   y = runif(999) - 0.5
   z = cos(2*pi*(x^2+y^2))
   ans = akimaInterp(x, y, z, extrap = FALSE)
   persp(ans)
   contour(ans)
   
## krigeInterp -
   # Kriging:    
   ans = krigeInterp(x, y, z, extrap = FALSE)
   persp(ans)
   contour(ans)

[Package fUtilities version 251.70 Index]