HolidayCalendars(fCalendar)	R Documentation

Holiday Calendars and Utilities

Description

A collection and description of functions and methods dealing with holiday dates and holiday calendars.

The functions for the generation of holiday calendars are:

easter 	 Returns date of easter or related feasts as 'sdate'-s,

holiday 	 Returns a holiday date of G7 and CH as 'sdate',

holiday.NYSE 	 Computes 'timeDate' object for full-day NYSE holidays.

SPlus like Functions are:

.julian 	 Converts formatted calendar dates to Julian counts,

month.day.year 	 Returns months, days, and years from Julian counts,

leap.year 	 Deciding if a year is a leap year or not,

day.of.week 	 Determines the day of week, starting with Sunday as 0.

Usage


easter(year, shift = 0)
holiday(year, Holiday = Easter)
holiday.NYSE(y = currentYear)

.julian(m, d, y, origin = c(month = 1, day = 1, year = 1960))
month.day.year(jul, origin = c(month = 1, day = 1, year = 1960))
leap.year(y) 
day.of.week(month, day, year)

Arguments

day 	 an integer vector of day numbers, [DD]. [day.of.week] - cr vector of integer day numbers.

Holiday 	 the unquoted function name of an ecclestial or public holiday in the G7 countries or Switzerland, see the list below.

jul 	 [month.day.year] -
vector of Julian dates (the number of days since origin).

m, d, y 	 [.julian] -
vector of month, day and year numbers. The years should be entered in their full four-digit form; it is assumed that two-digit years refer to the first century.

month 	 an integer vector of month numbers, [MM].
[day.of.week] - cr vector of integer month numbers.

origin 	 the origin date of the counter, in ISO-8601 date format, [CCYYMMDD]. By default Januar 1st, 1960.

shift 	 a numeric value, the n-day shift. By default the shift is set to zero and the function returns the date of Easter.

year 	 an integer vector of years, [CCYY]. [day.of.week] - cr vector of integer year numbers.

Details

Holiday Calendar:

Easter is the central ecclestial holiday. Many other holidays are related to this feast. The function easter computes the dates of easter and related ecclestial holidays for the requested year vector. holiday calculates the dates of ecclestial or publich holidays in the G7 countries, e.g. holiday(2003, "GoodFriday"). holidays.db is our Data Base of World Wide Holidays. The database contains holiday functions automatically loaded by the fBasics library at startup time. The user can add additional holiday functions to this data base. The information for the holiday data base were collected from several web pages about holiday calendars. The following ecclestial and public [HOLIDAY] functions in the G7 countries and Switzerland are available:

Holidays Related to Easter:
Septuagesima, Quinquagesima, AshWednesday, PalmSunday, GoodFriday, EasterSunday, Easter, EasterMonday, RogationSunday, Ascension, Pentecost, PentecostMonday, TrinitySunday CorpusChristi.

Holidays Related to Christmas:
ChristTheKing, Advent1st, Advent1st, Advent3rd, Advent4th, ChristmasEve, ChristmasDay, BoxingDay, NewYearsDay.

Other Ecclestical Feasts:
SolemnityOfMary, Epiphany, PresentationOfLord, Annunciation, TransfigurationOfLord, AssumptionOfMary, AssumptionOfMary, BirthOfVirginMary, CelebrationOfHolyCross, MassOfArchangels, AllSaints, AllSouls.

CHZurich - Public Holidays:
CHBerchtoldsDay, CHSechselaeuten, CHAscension, CHConfederationDay, CHKnabenschiessen.

GBLondon - Public Holidays:
GBMayDay, GBBankHoliday, GBSummerBankHoliday, GBNewYearsEve.

DEFrankfurt - Public Holidays:
DEAscension, DECorpusChristi, DEGermanUnity, DEChristmasEve, DENewYearsEve.

FRParis - Public Holidays:
FRFetDeLaVictoire1945, FRAscension, FRBastilleDay, FRAssumptionVirginMary, FRAllSaints, FRArmisticeDay.

ITMilano - Public Holidays:
ITEpiphany, ITLiberationDay, ITRepublicAnniversary, ITAssumptionOfVirginMary, ITAllSaints, ITWWIVictoryAnniversary, ITStAmrose, ITImmaculateConception.

USNewYork/USChicago - Public Holidays:
USNewYearsDay, USInaugurationDay, USMLKingsBirthday, USLincolnsBirthday, USWashingtonsBirthday, USMemorialDay, USIndependenceDay, USLaborDay, USColumbusDay, USElectionDay, USVeteransDay, USThanksgivingDay, USChristmasDay, USCPulaskisBirthday, USGoodFriday.

CAToronto/CAMontreal - Public Holidays:
CAVictoriaDay, CACanadaDay, CACivicProvincialHoliday, CALabourDay, CAThanksgivingDay, CaRemembranceDay.

JPTokyo/JPOsaka - Public Holidays:
JPNewYearsDay, JPGantan, JPBankHolidayJan2, JPBankHolidayJan3, JPComingOfAgeDay, JPSeijinNoHi, JPNatFoundationDay, JPKenkokuKinenNoHi, JPGreeneryDay, JPMidoriNoHi, JPConstitutionDay, JPKenpouKinenBi, JPNationHoliday, JPKokuminNoKyujitu, JPChildrensDay, JPKodomoNoHi, JPMarineDay, JPUmiNoHi, JPRespectForTheAgedDay, JPKeirouNoHi, JPAutumnalEquinox, JPShuubun-no-hi, JPHealthandSportsDay, JPTaiikuNoHi, JPNationalCultureDay, JPBunkaNoHi, JPThanksgivingDay, JPKinrouKanshaNohi, JPKinrou-kansha-no-hi, JPEmperorsBirthday, JPTennou-tanjyou-bi, JPTennou-tanjyou-bi.

All the holiday functions are listed in the data file holidays.R. Additional holidays, which are not yet available there, can be added to this data base file.

Value

easter
holiday
holiday.NYSE
return an ISO-8601 formatted calendar date CCYYMMDD, an object of class "sdate" which represents an integer value or vector.

leap.year
returns a logical vector indicating whether the input years are leap years or not.

day.of.week
returns a number between 0 and 6 that specifies the day of the week of the given date; 0 refers to Sunday.

Note

Most of the functions were written for R Version 1.5. With the new Date Class becoming available in R 1.9.0 these functions will undergo a revision in the near future.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

References

ISO-8601 (1988); Data Elements and Interchange Formats - Information Interchange, Representation of Dates and Time, International Organization for Standardization, Reference Number ISO 8601, 14 pages.

Examples


## SOURCE("fCalendar.5B-HolidayCalendars")

## easter -
   # Dates for Easter and Pentecost from 2000 until 2010:
   easter(2000:2010)
   easter(2000:2010, -2) # This is Good Friday
   holiday(2000:2010, "Easter")                      
   holiday(2000:2010, "GoodFriday")                  
   
## holidays -   
   Easter(2000:2010)                                
   GoodFriday(2000:2010)
  
## holiday.NYSE -
   holiday.NYSE(currentYear)
   

[Package fCalendar version 231.10066 Index]