fBasicsTools(fBasics)	R Documentation

fBasics Tools

Description

Tools used in the fBasics library.

Usage


xmpfBasics()
xmpBasics(prompt = "")

Arguments

prompt 	 the string printed when prompting the user for input.

Value

xmpfBasics
Popups the example menu.

xmpBasics
Nothing, the default, or the the prompt if you have set xmpBasics = readline on the command prompt.

Note

The example in the manual pages may be interactive and ask for input from the user. To achieve this you have to type on the command line: xmpBasics = readline

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fBasics.6A-BasicsTools")

[Package fBasics version 221.10066 Index]