GeneralS4Extractors(fUtilities)	R Documentation

General S4 Class Extractor Functions

Description

A collection and description of functions to extract slots from S4 class objects.

The extractor functions are:

isS4 	 Checks if an object is a S4 object,

getCall 	 Extracts the call slot from a S4 object,

getModel 	 Extracts the model slot from a S4 object,

getTitle 	 Extracts the title slot from a S4 object,

getDescription 	 Extracts the description slot from a S4 object,

getSlot 	 Extracts a specified slot from a S4 object.

Usage


isS4(object)

getCall(object)
getModel(object)
getTitle(object)
getDescription(object)

getSlot(object, slotName)

Arguments

object 	 an object of class S4.

slotName 	 [getSlot] -
a character string, the name of the slot to be extracted from the S4 object.

Author(s)

Diethelm Wuertz for this R port.

Examples


## Example S4 Representation:
   # Hyothesis Testing with Control Settings 
   setClass("hypTest", 
     representation(
       call = "call",
       data = "numeric",
       test = "list",
       description = "character")  
   )
   
## Shapiro Wilk Normaility Test
   swTest = function(x, description = "") {
     ans = shapiro.test(x)
     class(ans) = "list"
     new("hypTest", 
       call = match.call(), 
       data = x, 
       test = ans,
       description = description)
   }
   test = swTest(x = rnorm(500), description = "500 RVs")
   
## Extractor Functions:
   isS4(test)
   getCall(test)
   getDescription(test)

[Package fUtilities version 251.70 Index]