CalendarTools(fCalendar)	R Documentation

fCalendar Tools

Description

A collection and description of functions and tools used in the 'fCalendar' package. This includes the setting of the local environment and the management of examples.

Usage



xmpfCalendar()
xmpCalendar(prompt = "")


Arguments

prompt 	 the string printed when prompting the user for input.

Details

The function datax is an extended version of R's function data. It comes with the same functionality beside the fact, that data files loaded from the demo directories are transformed into S4 "timeSeries" objects if the data structure allows this.

For details for the function data we refer to the original help page.

Value

xmpfCalendar
Popups the example menu.

xmpCalendar
Nothing, the default, or the the prompt if you have set xmpCalendar = readline on the command prompt.

Note

The example in the manual pages may be interactive and ask for input from the user. To achieve this you have to type on the command line: xmpCalendar = readline

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCalendar.7A-CalendarTools")   

[Package fCalendar version 231.10066 Index]