ACD_Fit(fACD)	R Documentation

Function for Estimation of an ACD Model

Description

This function estimates, based on maximum likelihood, the parameters of an ACD (autoregressive conditional duration) model.

Usage

ACD_Fit(x, qLag = 1, pLag = 1, distrib = "exp", typeACD = "ACD")

Arguments

x 	 Vector with observed durations

qLag 	 Maximum lag for alpha parameter

pLag 	 Maximum lag for beta parameter

distrib 	 Assumed distribution ('exp' or 'weibull')

typeACD 	 Functional form of conditional duration equation (possible value = 'ACD','log','BC' or 'EX')

Value

Returns a S4 object with the following slots:

x 	 Observed Durations

qLag 	 Maximum lag for alpha parameter

pLag 	 Maximum lag for beta parameter

condDur 	 Conditional Duration Series

Coeff 	 A list with all estimated coefficients

Coeff_Std 	 A list with all standard errors for coefficients

Coeff_pValues 	 A list with all p values of coefficients

LL 	 Value of maximum log likelihood

paramVec 	 A vector with all coefficients (same values as in Coeff)

nParameter 	 Number of parameters in the model

sizeModel 	 A list with the size of the model (number of indep var, etc)

distrib 	 Assumed distribution for ML estimation

typeACD 	 Assumed functional form of ACD filter

timeRun 	 Time of estimation of model

Author(s)

Marcelo Perlin - ICMA/UK <marceloperlin@gmail.com>

See Also

ACD_Simul,acdModel-class

Examples



data(myDur)

x=myDur
pLag=1
qLag=1
distrib='exp'

typeACD='ACD'

modelOut<-ACD_Fit(x,qLag,pLag,distrib,typeACD)

[Package fACD version 1.0 Index]