ACD_Lik(fACD)	R Documentation

Function for calculation of log likelihood of ACD model

Description

This function calculates the conditional duration and resulting log likelihood of ACD models. This function is called by ACD_Fit.

Usage

ACD_Lik(param, x, qLag, pLag, distrib, typeACD, typeCall = "maxLik")

Arguments

param 	 parameter vector

x 	 Observed duration series

qLag 	 Maximum lag for alpha parameter

pLag 	 Maximum lag for beta parameter

distrib 	 Distribution assumption

typeACD 	 Type of ACD model

typeCall 	 Type of Call to function

Value

Returns the log likelihood of the model, or filtered series (dependeding on typeCall)

Author(s)

Marcelo Perlin - ICMA/UK <marceloperlin@gmail.com>

[Package fACD version 1.0 Index]