MS_Model-class(fMarkovSwitching)	R Documentation

Definitions for Class 'MS_Model'

Description

This is the class of a fitted Markov Switching Model S4 object. This is a output object from MS_Regress_Fit so normally you wont create MS_Model objects directly

Slots

filtProb:

Class:"matrix" - Filtered States Probabilities from the model (number of columns equal to k)

smoothProb:

Class:"matrix" - Smoothed States Probabilities from the model by means of Kim's Algorithm

Coeff:

Class:"list" - A list with all estimated coefficients

condMean:

Class:"matrix" - The fitted conditional Mean (based on t¦t)

condStd :

Class:"matrix" - The fitted conditional Standard Deviation (based on t¦t)

Coeff_Std:

Class:"list" - A list with the standard error of coefficients (same slots as @Coeff)

LL:

Class:"numeric" - The final log likelihood of the model

k:

Class:"numeric" - Number of States in the model

paramVec:

Class:"numeric" - The vector of parameters with the same values as Coeff, but in vector notation (you probably wont use it)

stateDur:

Class:"numeric" - The expected duration of each regime (how long it take to change states)

nParameter:

Class:"numeric" - Number of parameters in the model

sizeModel:

Class:"list" - A list with information regarding the size of the model

distrib:

Class:"character" - The assumed distribution (Normal or r)

Methods

plot

Plots an object of class 'MS_Model'

print

Prints to screen the fitted MS_Model

dim

Prints to screen the dimension/size of MS_Model

Author(s)

Marcelo Perlin - ICMA/UK <marceloperlin@gmail.com>

See Also

MS_Simul

Examples


showClass("MS_Model")

[Package fMarkovSwitching version 1.0 Index]