MS_Regress_Lik(fMarkovSwitching)	R Documentation

The likelihood function for a General Markov Switching Model

Description

This function calculates the log likelihood of the model and is maximized by Rdonlp2. You'll probably never have to call this function (MS_Regress_Fit does that for you..)

Usage


MS_Regress_Lik(param)

Arguments

param 	 a vector with parameters, only. All the rest of the input arguments are passed as global variables (see rdonlp2 documentation)

Value

Returns the sum of log likelihood. It may also return the filtered series of the model (see MS_Regress_Fit code)

Author(s)

Marcelo Perlin - ICMA/UK <marceloperlin@gmail.com>

References

ALEXANDER, C. (2008) 'Market Risk Analysis: Practical Financial Econometrics' Wiley
HAMILTON, J., D. (1994) 'Time Series Analysis' Princeton University Press
HAMILTON, J., D. (2005) ‘Regime Switching Models’ Palgrave Dictionary of Economic
KIM, C., J., NELSON, C., R. (1999) ‘State Space Model with Regime Switching: Classical and Gibbs-Sampling Approaches with Applications’ The MIT press

See Also

MS_Regress_Simul,MS_Regress_For,MS_Regress_Fit

Examples

# No Example. This function should normally not be called

[Package fMarkovSwitching version 1.0 Index]