MS_Simul-class(fMarkovSwitching)	R Documentation

A Class for the Simulation of a Markov Switching Time Series Process

Description

This class is the output from the use of function MS_Regress_Simul. Normally you'll never call the object contructor to build this class as this task is already done by MS_Regress_Simul

Slots

nr:

Class:"numeric" - Number of time periods simulated

dep:

Class:"matrix" - Simulated Explained/dependent Variable (a Markov Switching Process)

indep:

Class:"matrix" - Simulated explanatory variables (arbitrarily normal random numbers)

Coeff:

Class:"list" - A list with all coefficients inputed to the function MS_Regress_Simul

S:

Class:"numeric" - The input that controls where to switch (from the call to MS_regress_Simul

trueStates:

Class "matrix" - The "true" states of the model which where build according to transation matrix (Coeff$P)

k:

Class:"numeric" - Number of states in the model

distrib:

Class:"character" - Distribution assumption for residue

Methods

plot

Plots an object of class 'MS_Simul'

print

Prints to screen the simulated model (the 'MS_Simul' object)

Author(s)

Marcelo Perlin - ICMA/UK <marceloperlin@gmail.com>

See Also

MS_Model for links to other classes

Examples


showClass("MS_Simul")

[Package fMarkovSwitching version 1.0 Index]