SkewHyperbolic-package(SkewHyperbolic)	R Documentation

The Package 'SkewHyperbolic': Summary Information

Description

This package provides a collection of functions for working with the skew hyperbolic Student's t-distribution.

Functions are provided for the density function (dskewhyp), distribution function (pskewhyp), quantiles (qskewhyp) and random number generation (rskewhyp). There are also functions that fit the distribution to data (skewhypFit). The mean, variance, skewness, kurtosis and mode can be found usig the functions skewhypMean, skewhypVar, skewhypSkew, skewhypKurt and skewhypMode respectively. To assess goodness of fit, there are also functions to generate a Q-Q plot (qqskewhyp) and a P-P plot (ppskewhyp). S3 methods print, plot and summary are provided for the output of skwewhypFit.

Package: 	 SkewHyperbolic

Type: 	 Package

Version: 	 0.1-1

Date: 	 2009-08-28

License: 	 GPL(>=2)

LazyLoad: 	 yes

Author(s)

David Scott d.scott@auckland.ac.nz, Fiona Grimson

References

Aas, K. and Haff, I. H. (2006). The Generalised Hyperbolic Skew Student's t-distribution, Journal of Financial Econometrics, 4, 275–309.

See Also

dskewhyp, skewhypMean, skewhypFit, skewhypFitStart, skewhypBreaks, qqskewhyp, HyperbolicDist.

[Package SkewHyperbolic version 0.1-1 Index]