TermStructure(fBonds)	R Documentation

Term Structure Modelling

Description

A collection and description of functions for term structure modelling.

The functions are:

NelsonSiegel 	 Nelson-Siegel Term Structure,

Svensson 	 Nelson-Siegel-Svensson Term Structure.

Usage


NelsonSiegel(rate, maturity, doplot = TRUE)
Svensson(rate, maturity, doplot = TRUE)

Arguments

doplot 	 a logical. Should a plot be displayed?

maturity 	 a numeric vector maturities on an annual scale.

rate 	 a numeric vectore of forward rates.

Value

a list object with entries returned from the optimization function nlminb.

References

McCulloch J. H. (1990); US Term Structure Data: 1946-87, Handbook of Monetary Economics, Friedman B.M. and Hahn F.H. (eds.), Elsevier Science.

McCulloch J. H. and Kwon, H.C. (1993); US Term Structure Data: 1947-1991, Working Paper No. 93-6, Department of Economics, Ohio State University.

Zivot E., Wang J.; Modeling Financial Time Series with S-Plus.

[Package fBonds version 221.10065 Index]