acdModel-class(fACD)	R Documentation

A Class for an ACD Model

Description

This class is the output from the use of function ACD_Fit. Normally you'll never call the object contructor to build this class

Slots

x

Observed Durations

qLag

Maximum lag for alpha parameter

pLag

Maximum lag for beta parameter

condDur

Conditional Duration Series

Coeff

A list with all estimated coefficients

Coeff_Std

A list with all standard errors for coefficients

Coeff_pValues

A list with all p values of coefficients

LL

Value of maximum log likelihood

paramVec

A vector with all coefficients (same values as in Coeff)

nParameter

Number of parameters in the model

sizeModel

A list with the size of the model (number of indep var, etc)

distrib

Assumed distribution for ML estimation

typeACD

Assumed functional form of ACD filter

timeRun

Time of estimation of model

Methods

plot

Plots an object of class 'acdModel'

print

Prints to screen the coefficients of the model (the 'acdModel' object)

Author(s)

Marcelo Perlin - ICMA/UK <marceloperlin@gmail.com>

Examples

showClass("acdModel")

[Package fACD version 1.0 Index]