dep(fMarkovSwitching)	R Documentation

Example of dependent/explained Variable for Markov switching model

Description

Just an example data, in this case logarithm returns. This example data is the same one from the MatLab version of the package

Usage

data(dep)

Format

The format is: num [1:1000] -0.07318 0.05899 0.00795 0.00000 0.00717 ...

Examples


data(dep)
## maybe str(dep) ; plot(dep) ...

[Package fMarkovSwitching version 1.0 Index]