dim.MS_Model(fMarkovSwitching)	R Documentation

Function to Get the dimension/Size of Markov Switching Variables

Description

This function will output the dimensions of the model, that is, how many observations, how many parameters, how many states, etc..

Usage


dim.MS_Model(MS_Model_In)

Arguments

MS_Model_In 	 A S4 Object obtained with MS_Regress_Fit

Value

Returns nothing. Prints the output in R screen

Author(s)

Marcelo Perlin - ICMA Centre/UK

See Also

MS_Regress_Fit,MS_Regress_Simul,MS_Regress_For

[Package fMarkovSwitching version 1.0 Index]