fBondsData(fBonds)	R Documentation

fBonds Data Sets

Description

A collection and description of data sets from the Rmetrics Package fBonds.

The data sets are:

bundesbankNNS 	 Nelson-Siegel-Svensson Coefficients.

Details

Bundesbank Nelson-Siegel-Svensson Coefficients:

Coefficients for the Nelson-Siegel-Svensson yield curve from the German Bundesbank.

:

mk.zero2 is a data set with a 67 x 55 values representing the zero-coupon yield curve from August 1985 to February 1991.
mk.maturity a numeric vector of length 55, giving the fifty-five maturities in terms of years for the term structure.

References

McCulloch J. H. (1990); US term structure data: 1946-87, Handbook of Monetary Economics, Friedman B.M. and Hahn F.H. (eds.), Elsevier Science.

McCulloch J. H. and Kwon, H.C. (1993); US term structure data: 1947-1991, Working Paper No. 93-6, Department of Economics, Ohio State University.

[Package fBonds version 221.10065 Index]