fMarkovSwitching-package(fMarkovSwitching)	R Documentation

Package for Estimation Simulation and Forecasting of a Markov Switching Model

Description

This package has designed the estimation, simulation and forecasting of Markov Switching models. See examples for simple ways of using the package and also check the pdf document for advanced examples of functions

Details

Package: 	 fMarkovSwitching

Type: 	 Package

Version: 	 1.0

Date: 	 2008-10-27

License: 	 GPL (>=2)

LazyLoad: 	 yes

Author(s)

Marcelo Perlin - ICMA/UK <marceloperlin@gmail.com>

References

ALEXANDER, C. (2008) 'Market Risk Analysis: Practical Financial Econometrics' Wiley
HAMILTON, J., D. (1994) 'Time Series Analysis' Princeton University Press
HAMILTON, J., D. (2005) 'Regime Switching Models' Palgrave Dictionary of Economic
KIM, C., J., NELSON, C., R. (1999) 'State Space Model with Regime Switching: Classical and Gibbs-Sampling Approaches with Applications' The MIT press

Examples

#See each function example (or pdf document)..

[Package fMarkovSwitching version 1.0 Index]