indep(fMarkovSwitching)	R Documentation

Example of independent/explanatory Variable for Markov switching model

Description

Just an example data, in this case logarithm returns. This example data is the same one from the MatLab version of the package

Usage

data(indep)

Format

The format is: num [1:1000, 1:3]

Details

The data consists of one vector or ones (the constant) and 2 vectors of time series

Examples

# See MS_Regress_Fit for example

[Package fMarkovSwitching version 1.0 Index]