lrdji(SkewHyperbolic)	R Documentation

Dow Jones Log Return Data

Description

Log returns of daily closing value data from the dow jones index, from 04/JAN/1999 to 08/JUL/2003. The original data used to calculate these was the dji data set available in the QRMlib package.

Usage


data(lrdji)

Format

A vector of 1132 observations.

Author(s)

David Scott d.scott@auckland.ac.nz, Fiona Grimson

Source

library(QRMlib) data(dji)

References

McNeil, A. & Ulman, S. (2008). QRMlib http://cran.r-project.org/web/packages/QRMlib/index.html

Examples


data(lrdji)
##fit a skew hyperbolic students t-distribution to the data
fit<-skewhypFit(lrdji, plot=TRUE, print=TRUE)

[Package SkewHyperbolic version 0.1-1 Index]