summary.vgFit(VarianceGamma)	R Documentation

Summarizing Variance Gamme Distribution Fit

Description

summary Method for class "vgFit".

Usage


## S3 method for class 'vgFit':
summary(object, ...)

## S3 method for class 'summary.vgFit':
print(x, digits = max(3, getOption("digits") - 3), ...)

Arguments

object 	 An object of class "vgFit", resulting from a call to vgFit.

x 	 An object of class "summary.vgFit", resulting from a call to summary.vgFit.

digits 	 The number of significant digits to use when printing.

... 	 Further arguments passed to or from other methods.

Details

summary.vgFit calculates standard errors for the estimates of c, sigma, theta, and nu of the variance gamma distribution parameter vector param if the Hessian from the call to optim or nlm is available. Because the parameters in the call to the optimiser are c, log(sigma), theta and log(nu), the delta method is used to obtain the standard errors for sigma and nu.

Value

If the Hessian is available, summary.vgFit computes standard errors for the estimates of c, sigma, theta, and nu, and adds them to object as object$sds. Otherwise, no calculations are performed and the composition of object is unaltered.
summary.vgFit invisibly returns x with class changed to summary.vgFit.
See vgFit for the composition of an object of class vgFit.
print.summary.vgFit prints a summary in the same format as print.vgFit when the Hessian is not available from the fit. When the Hessian is available, the standard errors for the parameter estimates are printed in parentheses beneath the parameter estimates, in the manner of fitdistr in the package MASS.

See Also

vgFit, summary.

Examples


### Continuing the  vgFit(.) example:
param <- c(0,0.5,0,0.5)
dataVector <- rvg(500, param = param)
fit <- vgFit(dataVector)
print(fit)
summary(fit)          

[Package VarianceGamma version 0.1-1 Index]