vgChangePars(VarianceGamma)	R Documentation

Change Parameterizations of the Variance Gamma Distribution

Description

This function interchanges between the following 4 parameterizations of the variance gamma distribution:

1. c, sigma, theta, nu

2. theta, sigma, mu, tau

3. theta, sigma, kappa, tau

4. lambda, alpha, beta, mu

The first set of parameterizations is given in Seneta (2004). The second and third ones are the parameterizations given in Kotz et al. (2001). The last set takes the form of the generalized hyperbolic distribution parameterization. delta is not included since the variance gamma distribution is a limiting case of generalized hyperbolic distribution with delta always equal to 0.

Usage


vgChangePars(from, to, param, noNames = FALSE)

Arguments

from 	 The set of parameters to change from.

to 	 The set of parameters to change to.

param 	 "from" parameter vector consisting of 4 numerical elements.

noNames 	 Logical. When TRUE, suppresses the parameter names in the output.

Details

In the 3 parameterizations, the following must be positive:

1. sigma, nu

2. sigma, tau

3. sigma, tau

4. lambda, alpha

In addition in the 4th parameterization, the absolute value of beta must be less than alpha.

Value

A numerical vector of length 4 representing param in the to parameterization.

Author(s)

David Scott d.scott@auckland.ac.nz, Christine Yang Dong c.dong@auckland.ac.nz

References

Seneta, E. (2004). Fitting the variance-gamma model to financial data. J. Appl. Prob., 41A:177–187. Kotz, S, Kozubowski, T. J., and Podgórski, K. (2001). The Laplace Distribution and Generalizations. Birkhauser, Boston, 349 p.

See Also

dvg, code{vgMom}

Examples


param1 <- c(2,2,1,3)                   # Parameterization 1
param2 <- vgChangePars(1, 2, param1)   # Convert to parameterization 2
param2                                 # Parameterization 2
vgChangePars(2, 1, as.numeric(param2)) # Convert back to parameterization 1

param3 <- c(1,2,0,0.5)                 # Parameterization 3
param1 <- vgChangePars(3, 1, param3)   # Convert to parameterization 1
param1                                 # Parameterization 1
vgChangePars(1, 3, as.numeric(param1)) # Convert back to parameterization 3

[Package VarianceGamma version 0.2-1 Index]