vgFitStart(VarianceGamma)	R Documentation

Find Starting Values for Fitting a Variance Gamma Distribution

Description

Finds starting values for input to a maximum likelihood routine for fitting variance gamma distribution to data.

Usage


  vgFitStart(x, breaks = NULL, startValues = "SL", paramStart = NULL,
             startMethodSL = "Nelder-Mead",
             startMethodMoM = "Nelder-Mead", ...)
  vgFitStartMoM(x, startMethodMoM = "Nelder-Mead", ...)

Arguments

x 	 Data vector.

breaks 	 Breaks for histogram. If missing, defaults to those generated by hist(x, right = FALSE, plot = FALSE).

startValues 	 Vector of the different starting values to consider. See Details.

paramStart 	 Starting values for param if startValues = "US".

startMethodSL 	 Method used by call to optim in finding skew Laplace estimates.

startMethodMoM 	 Method used by call to optim in finding method of moments estimates.

... 	 Passes arguments to optim.

Details

Possible values of the argument startValues are the following:

"US"

User-supplied.

"SL"

Based on a fitted skew-Laplace distribution.

"MoM"

Method of moments.

If startValues = "US" then a value must be supplied for paramStart.

If startValues = "MoM", vgFitStartMoM is called. These starting values are based on Barndorff-Nielsen et al (1985).

If startValues = "SL", or startValues = "MoM" an initial optimisation is needed to find the starting values. These optimisations call optim.

Value

vgFitStart returns a list with components:

vgStart 	 A vector with elements vgC, lSigma (log of sigma), theta and lNu (log of nu) giving the starting value of param.

xName 	 A character string with the actual x argument name.

breaks 	 The cell boundaries found by a call to hist.

midpoints 	 The cell midpoints found by a call to hist.

empDens 	 The estimated density found by a call to hist.

vgFitStartMoM returns only the method of moments estimates as a vector with elements vgC, lSigma (log of sigma), theta and lNu (log of nu).

Author(s)

David Scott d.scott@auckland.ac.nz, Christine Yang Dong c.dong@auckland.ac.nz

References

Seneta, E. (2004). Fitting the variance-gamma model to financial data. J. Appl. Prob., 41A:177–187.

See Also

dvg, dskewlap, vgFit, hist, and optim.

Examples


param <- c(0,0.5,0,0.5)
dataVector <- rvg(500, param = param)
vgFitStart(dataVector,startValues="SL")
vgFitStartMoM(dataVector)
vgFitStart(dataVector,startValues="MoM")

[Package VarianceGamma version 0.2-1 Index]