ExtremeValueCopulae(fCopulae)	R Documentation

Bivariate Extreme Value Copulae

Description

A collection and description of functions to investigate bivariate extreme value copulae.

Extreme Value Copulae Functions:

revCopula 	 Generates extreme value copula random variates,

pevCopula 	 computes extreme value copula probability,

devCopula 	 computes extreme value copula density,

revSlider 	 displays interactive plots of extreme value random variates,

pevSlider 	 displays interactive plots of extreme value probability,

devSlider 	 displays interactive plots of extreme value density,

Afunc 	 computes dependence function,

AfuncSlider 	 displays interactively dependence function,

evTau 	 computes Kendall's tau for extreme value copulae,

evRho 	 computes Spearman's rho for extreme value copulae,

evTailCoeff 	 computes tail dependence for extreme value copulae,

evTailPlot 	 plots tail dependence for extreme value copulae,

evCopulaSim 	 simulates an extreme value copula,

evCopulaFit 	 fits the parameters of an extreme value copula.

Usage



revCopula(n, param = NULL, type = c("gumbel", "galambos", "husler.reiss", 
    "tawn", "bb5"))
pevCopula(u = 0.5, v = u, param = NULL, type = c("gumbel", "galambos", 
    "husler.reiss", "tawn", "bb5"), output = c("vector", "list"), 
    alternative = FALSE )
devCopula(u = 0.5, v = u, param = NULL, type = c("gumbel", "galambos", 
    "husler.reiss", "tawn", "bb5"), output = c("vector", "list"), 
    alternative = FALSE )   

revSlider(B = 10)
pevSlider(type = c("persp", "contour"), B = 10)
devSlider(type = c("persp", "contour"), B = 10)

evTau(param = NULL, type = 1:22, lower = 1.0e-10)
evRho(param = NULL, type = 1:22, method = c("integrate2d", "adapt"), error = 1.0e-5)
    
evTailCoeff(param = NULL, type = 1:22)
evTailPlot(param = NULL, type = 1:22, tail = c("Upper", "Lower"))

Afunc(x, param = NULL, type = c("gumbel", "galambos", "husler.reiss", "tawn", "bb5") )
AfuncSlider()

evCopulaSim(n, param = NULL,  type = c("gumbel", "galambos", "husler.reiss", 
    "tawn", "bb5"))
evCopulaFit(u, v = NULL, type = c("gumbel", "galambos", "husler.reiss", 
    "tawn", "bb5"), ...)

Arguments

alternative 	 [*evCopula] -
Should the probability be computed alternatively in a direct way from the probability formula or by default via the dependency function?

B 	 [*Slider] -
the maximum slider menu value when the boundary value is infinite. By default this is set to 10.

error 	 [evRho] -
the error bound to be achieved by the integrate2d integration formula. A numeric value, by default error=1.0e-5.

lower 	 [evTau] -
a numeric value setting the lower bound for the internal integration function integrate.

n 	 [revCopula][evCopulaSim] -
the number of random deviates to be generated, an integer value.

method 	 [evRho] -
a character string that determines which integration should be used, either "integrate2d" or "adapt". If the second method is selected the condtributed R package "adapt" is required.

output 	 [*evCopula] -
output - a character string specifying how the output should be formatted. By default a vector of the same length as u and v. If specified as "list" then u and v are expected to span a two-dimensional grid as outputted by the function grid2d and the function returns a list with elements $x, y, and z which can be directly used for example by 2D plotting functions.

param 	 [*ev*][A*] -
distribution and copulae parameters. A numeric value or vector of named parameters as required by the copula specified by the variable type. If set to NULL, then the default parameters will be taken.

tail 	 [evTailPlot] -
a character string, either "Upper" or "Lower" denoting which of the two tails should be displayed. By default the upper tail dependence will be considered.

type 	 [*ev*][Afunc] -
the type of the maximum extreme value copula. A character string selected from: "gumbel", "galambos", "husler.reiss", "tawn", or "bb5".
[evSlider] -
a character string specifying the plot type. Either a perspective plot which is the default or a contour plot with an underlying image plot will be created.

u, v 	 [*evCopula][*archmCopula] -
two numeric values or vectors of the same length at which the copula will be computed. If u is a list then the the $x and $y elements will be used as u and v. If u is a two column matrix then the first column will be used as u and the the second as v.

x 	 [A*] -
a numeric value or vector ranging between zero and one.

... 	 [evCopulaFit] -
arguments passed to the optimization function nlminb.

Value

Th functions *Spec return an S4 object of class "fCOPULA". The object contains the following slots:

@call 	 the function call.

@copula 	 the name of the copula.

@param 	 a list whose elements specify the model parameters.

@title 	 a character string with the name of the copula. This can be overwritten specifying a user defined input argument.

@description 	 a character string with an optional user defined description. By default just the current date when the test was applied will be returned.

The function pcopula returns a numeric matrix of probabilities computed at grid positions x|y.

The function parchmCopula returns a numeric matrix with values computed for the Archemedean copula.

The function darchmCopula returns a numeric matrix with values computed for thedensity of the Archemedean copula.

The functions Phi* return a numeric vector with the values computed from the Archemedean generator, its derivatives, or its inverse.

The functions cK and {cKInv} return a numeric vector with the values of the density and inverse for Archimedian copulae.

Author(s)

Diethelm Wuertz for the Rmetrics R-port.

Examples


## SOURCE("fCopulae.064A-ExtremeValueCopulae")
## Not run: 
## fCOPULA -
   # getClass("fCOPULA")
   
## pcopula -
   # The default Normal Copula:
   # contour(pcopula())
## End(Not run)

[Package fCopulae version 221.10065 Index]