vgParam(VarianceGamma)	R Documentation

Parameter Sets for Variance Gamma Distribution

Description

These objects store different parameter sets of the Variance Gamma distribution for testing or demonstrating purpose as matrixes. Specifically, the parameter sets vgSmallShape and vgLargeShape have constant (standard) location and spread parameters of c=0 and sigma=1; where asymmetry and shape parameters vary from theta=(-2, 0, 2) and nu=(0.5, 1, 2) for vgSmallShape and theta=(-4, -2, 0, 2, 4) and nu=(0.25, 0.5, 1, 2, 4) for vgLargeShape.

The parameter sets vgSmallParam and vgLargeParam have varied values of all 4 parameters. vgSmallParam contains all of the parameter combinations from c=(-2, 0, 2), sigma=(0.5, 1, 2), theta=(-2, 0, 2) and nu=(0.5, 1, 2). vgLargeParam contains all of the parameter combinations from c=(-4, -2, 0, 2, 4), sigma=(0.25, 0.5, 1, 2, 4), theta=(-4, -2, 0, 2, 4) and nu=(0.25, 0.5, 1, 2, 4).

Usage


  vgSmallShape
  vgLargeShape
  vgSmallParam
  vgLargeParam

Format

vgSmallShape: a 9 by 4 matrix; vgLargeShape: a 25 by 4 matrix; vgSmallParam: a 81 by 4 matrix; vgLargeParam: a 625 by 4 matrix.

Author(s)

David Scott d.scott@auckland.ac.nz, Christine Yang Dong c.dong@auckland.ac.nz

Examples


data(vgParam)
## Testing the accuracy of vgMean
for (i in 1:nrow(vgSmallParam)) {
    param <- vgSmallParam[i,]
    x <- rvg(10000,param = param)
    sampleMean <- mean(x)
    funMean <- vgMean(param = param)
    difference <- abs(sampleMean - funMean)
    print(difference)
}


[Package VarianceGamma version 0.2-1 Index]